Get the fitted values from a DFA as a data frame

Get the fitted values from a DFA as a data frame

dfa_fitted(modelfit, conf_level = 0.95, names = NULL)

Arguments

modelfit

Output from fit_dfa.

conf_level

Probability level for CI.

names

Optional vector of names for time series labels. Should be same length as the number of time series.

Value

A data frame with the following columns: ID is an identifier for each time series, time is the time step, y is the observed values standardized to mean 0 and unit variance, estimate is the mean fitted value, lower is the lower CI, and upper is the upper CI.

See also

Examples

# \donttest{
y <- sim_dfa(num_trends = 2, num_years = 20, num_ts = 4)
m <- fit_dfa(y = y$y_sim, num_trends = 2, iter = 50, chains = 1)
#> 
#> SAMPLING FOR MODEL 'dfa' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000541 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 5.41 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: WARNING: There aren't enough warmup iterations to fit the
#> Chain 1:          three stages of adaptation as currently configured.
#> Chain 1:          Reducing each adaptation stage to 15%/75%/10% of
#> Chain 1:          the given number of warmup iterations:
#> Chain 1:            init_buffer = 3
#> Chain 1:            adapt_window = 20
#> Chain 1:            term_buffer = 2
#> Chain 1: 
#> Chain 1: Iteration:  1 / 50 [  2%]  (Warmup)
#> Chain 1: Iteration:  5 / 50 [ 10%]  (Warmup)
#> Chain 1: Iteration: 10 / 50 [ 20%]  (Warmup)
#> Chain 1: Iteration: 15 / 50 [ 30%]  (Warmup)
#> Chain 1: Iteration: 20 / 50 [ 40%]  (Warmup)
#> Chain 1: Iteration: 25 / 50 [ 50%]  (Warmup)
#> Chain 1: Iteration: 26 / 50 [ 52%]  (Sampling)
#> Chain 1: Iteration: 30 / 50 [ 60%]  (Sampling)
#> Chain 1: Iteration: 35 / 50 [ 70%]  (Sampling)
#> Chain 1: Iteration: 40 / 50 [ 80%]  (Sampling)
#> Chain 1: Iteration: 45 / 50 [ 90%]  (Sampling)
#> Chain 1: Iteration: 50 / 50 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.077009 seconds (Warm-up)
#> Chain 1:                0.061353 seconds (Sampling)
#> Chain 1:                0.138362 seconds (Total)
#> Chain 1: 
#> Warning: There were 1 chains where the estimated Bayesian Fraction of Missing Information was low. See
#> https://mc-stan.org/misc/warnings.html#bfmi-low
#> Warning: Examine the pairs() plot to diagnose sampling problems
#> Warning: The largest R-hat is NA, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#r-hat
#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#bulk-ess
#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#tail-ess
#> Inference for the input samples (1 chains: each with iter = 25; warmup = 12):
#> 
#>                   Q5      Q50      Q95     Mean     SD  Rhat Bulk_ESS Tail_ESS
#> x[1,1]           1.6      1.6      1.6      1.6    0.0  2.06        4       13
#> x[2,1]           1.3      1.3      1.3      1.3    0.0  2.06        4       13
#> x[1,2]           0.6      0.6      0.6      0.6    0.0  2.06        3       13
#> x[2,2]           0.1      0.1      0.1      0.1    0.0  2.06        4       13
#> x[1,3]           0.1      0.2      0.2      0.2    0.0  2.06        4       13
#> x[2,3]          -1.5     -1.5     -1.5     -1.5    0.0  2.06        3       13
#> x[1,4]           0.8      0.8      0.8      0.8    0.0  2.06        4       13
#> x[2,4]          -0.5     -0.5     -0.5     -0.5    0.0  2.06        3       13
#> x[1,5]          -0.6     -0.6     -0.6     -0.6    0.0  2.06        4       13
#> x[2,5]           1.2      1.2      1.2      1.2    0.0  2.06        3       13
#> x[1,6]          -0.5     -0.5     -0.5     -0.5    0.0  1.10        9       13
#> x[2,6]          -0.5     -0.5     -0.5     -0.5    0.0  2.06        3       13
#> x[1,7]          -0.1     -0.1     -0.1     -0.1    0.0  0.95       13       13
#> x[2,7]          -0.1     -0.1     -0.1     -0.1    0.0  2.06        3       13
#> x[1,8]           0.5      0.5      0.5      0.5    0.0  1.13        9       13
#> x[2,8]           1.1      1.1      1.1      1.1    0.0  2.06        4       13
#> x[1,9]          -1.3     -1.3     -1.3     -1.3    0.0  1.24        8       13
#> x[2,9]           0.3      0.3      0.3      0.3    0.0  2.06        3       13
#> x[1,10]          0.7      0.7      0.7      0.7    0.0  0.94       12       13
#> x[2,10]         -1.3     -1.3     -1.3     -1.3    0.0  2.06        3       13
#> x[1,11]         -0.9     -0.9     -0.9     -0.9    0.0  1.45       13       13
#> x[2,11]         -1.2     -1.2     -1.2     -1.2    0.0  2.06        3       13
#> x[1,12]         -1.8     -1.8     -1.8     -1.8    0.0  1.09       11       13
#> x[2,12]         -0.6     -0.6     -0.6     -0.6    0.0  2.06        3       13
#> x[1,13]         -0.1     -0.1     -0.1     -0.1    0.0  2.06        4       13
#> x[2,13]         -2.1     -2.0     -2.0     -2.0    0.0  2.06        3       13
#> x[1,14]          0.8      0.8      0.8      0.8    0.0  2.06        3       13
#> x[2,14]         -1.4     -1.4     -1.4     -1.4    0.0  2.06        3       13
#> x[1,15]         -0.5     -0.5     -0.5     -0.5    0.0  2.06        4       13
#> x[2,15]         -0.8     -0.8     -0.7     -0.7    0.0  1.71        4       13
#> x[1,16]          1.6      1.6      1.6      1.6    0.0  1.58        4       13
#> x[2,16]          0.9      0.9      0.9      0.9    0.0  1.71        4       13
#> x[1,17]          0.7      0.7      0.7      0.7    0.0  1.19       13       13
#> x[2,17]          1.4      1.4      1.4      1.4    0.0  2.06        4       13
#> x[1,18]          0.3      0.3      0.3      0.3    0.0  1.87        4       13
#> x[2,18]          2.3      2.3      2.3      2.3    0.0  2.06        3       13
#> x[1,19]         -0.9     -0.9     -0.9     -0.9    0.0  1.87        4       13
#> x[2,19]          0.8      0.9      0.9      0.9    0.0  2.06        3       13
#> x[1,20]          0.3      0.3      0.3      0.3    0.0  1.87        4       13
#> x[2,20]          0.2      0.2      0.2      0.2    0.0  2.06        3       13
#> Z[1,1]         -97.8    -97.8    -97.8    -97.8    0.0  1.71        4       13
#> Z[2,1]          49.5     49.6     49.7     49.6    0.1  2.06        4       13
#> Z[3,1]         -58.8    -58.8    -58.7    -58.8    0.1  2.06        3       13
#> Z[4,1]          24.6     24.7     24.7     24.7    0.0  2.06       11       13
#> Z[1,2]           0.0      0.0      0.0      0.0    0.0  1.00       13       13
#> Z[2,2]         -99.6    -99.6    -99.6    -99.6    0.0  2.06        4       13
#> Z[3,2]          17.6     17.6     17.7     17.6    0.0  2.06        4       13
#> Z[4,2]         -66.7    -66.7    -66.5    -66.6    0.1  2.06        3       13
#> log_lik[1]   -2946.7  -2795.4  -2450.5  -2736.3  197.0  2.06        3       13
#> log_lik[2]    -338.2   -324.3   -289.3   -317.9   19.4  2.06        3       13
#> log_lik[3]    -564.2   -535.9   -466.9   -523.5   38.9  2.06        3       13
#> log_lik[4]    -321.8   -305.0   -268.5   -299.0   21.0  2.06        3       13
#> log_lik[5]    -453.7   -429.0   -373.1   -419.5   31.9  2.06        3       13
#> log_lik[6]     -75.9    -71.0    -58.8    -68.7    6.8  2.06        3       13
#> log_lik[7]    -138.0   -130.3   -112.2   -127.1   10.3  2.06        3       13
#> log_lik[8]     -13.4    -12.6    -10.4    -12.2    1.2  2.06        3       13
#> log_lik[9]     -29.0    -26.8    -22.3    -26.2    2.7  2.06        3       13
#> log_lik[10]  -3016.6  -2854.2  -2471.9  -2787.0  216.4  2.06        3       13
#> log_lik[11]   -147.4   -137.6   -117.4   -134.4   11.8  2.06        3       13
#> log_lik[12]  -1298.2  -1225.7  -1057.3  -1196.4   95.6  2.06        3       13
#> log_lik[13]   -812.5   -767.8   -671.4   -752.3   56.0  2.06        3       13
#> log_lik[14]  -1040.7   -979.7   -838.3   -955.2   80.4  2.06        3       13
#> log_lik[15]   -409.0   -385.4   -333.7   -376.9   29.8  2.06        3       13
#> log_lik[16]   -361.3   -340.1   -289.5   -331.1   28.6  2.06        3       13
#> log_lik[17]   -347.4   -332.6   -293.5   -325.1   21.6  2.06        3       13
#> log_lik[18]  -2703.5  -2579.9  -2291.3  -2529.4  163.8  2.06        3       13
#> log_lik[19]   -351.2   -334.3   -294.8   -327.3   22.3  2.06        3       13
#> log_lik[20]  -1097.6  -1045.6   -925.3  -1024.7   68.4  2.06        3       13
#> log_lik[21]   -345.5   -329.4   -289.7   -322.1   22.3  2.06        3       13
#> log_lik[22]    -91.9    -85.3    -70.3    -82.6    8.5  2.06        3       13
#> log_lik[23]    -60.5    -58.4    -52.1    -57.1    3.4  2.06        3       13
#> log_lik[24]    -65.5    -60.7    -50.4    -59.0    5.9  2.06        3       13
#> log_lik[25]    -17.7    -17.1    -15.1    -16.7    1.1  2.06        4       13
#> log_lik[26]     -7.7     -6.8     -5.0     -6.5    1.0  2.06        3       13
#> log_lik[27]     -3.4     -3.4     -3.4     -3.4    0.0  1.04       13       13
#> log_lik[28]     -5.5     -5.0     -3.9     -4.8    0.6  2.06        3       13
#> log_lik[29]   -315.7   -299.3   -266.2   -294.2   19.5  2.06        3       13
#> log_lik[30]   -771.1   -740.6   -665.4   -727.6   42.1  2.06        3       13
#> log_lik[31]    -18.3    -17.4    -15.5    -17.1    1.1  2.06        3       13
#> log_lik[32]   -406.0   -387.9   -346.3   -381.0   23.7  2.06        3       13
#> log_lik[33]  -1883.0  -1795.3  -1569.8  -1752.7  125.2  2.06        3       13
#> log_lik[34]  -1026.5   -986.4   -882.5   -967.3   57.4  2.06        3       13
#> log_lik[35]   -773.9   -738.2   -646.0   -720.7   51.1  2.06        3       13
#> log_lik[36]   -311.8   -300.8   -270.1   -294.9   16.7  2.06        3       13
#> log_lik[37]   -639.9   -603.7   -533.4   -593.5   41.6  2.06        3       13
#> log_lik[38]  -3314.8  -3124.6  -2695.2  -3050.0  246.2  2.06        3       13
#> log_lik[39]   -542.9   -510.1   -444.3   -500.0   38.6  2.06        3       13
#> log_lik[40]  -1332.2  -1253.9  -1076.0  -1222.8  101.9  2.06        3       13
#> log_lik[41]   -927.3   -886.1   -773.1   -863.9   62.2  2.06        3       13
#> log_lik[42]   -738.5   -690.5   -583.3   -671.8   61.8  2.06        3       13
#> log_lik[43]   -106.5   -103.5    -91.5   -100.8    6.2  2.06        3       13
#> log_lik[44]   -451.8   -421.5   -355.5   -410.2   38.2  2.06        3       13
#> log_lik[45]  -3669.6  -3499.0  -3066.5  -3418.2  241.6  2.06        3       13
#> log_lik[46]    -60.4    -59.6    -59.0    -59.6    0.5  1.02       11       13
#> log_lik[47]  -1007.6   -963.5   -845.5   -940.8   65.1  2.06        3       13
#> log_lik[48]     -1.8     -1.7     -1.7     -1.8    0.0  2.06        3       13
#> log_lik[49]    -19.7    -19.6    -17.8    -19.1    0.8  1.39        5       13
#> log_lik[50]  -4751.2  -4483.8  -3874.9  -4377.6  348.0  2.06        3       13
#> log_lik[51]   -101.2    -93.1    -79.8    -91.5    8.3  2.06        3       13
#> log_lik[52]  -2183.4  -2055.7  -1768.6  -2006.1  164.6  2.06        3       13
#> log_lik[53]   -694.9   -653.4   -570.8   -640.9   48.6  2.06        3       13
#> log_lik[54]  -3995.0  -3775.1  -3264.5  -3684.2  290.0  2.06        3       13
#> log_lik[55]   -640.6   -601.6   -521.3   -588.8   46.8  2.06        3       13
#> log_lik[56]  -1616.0  -1525.3  -1313.3  -1487.4  120.3  2.06        3       13
#> log_lik[57]   -284.1   -274.0   -241.2   -267.0   17.5  2.06        3       13
#> log_lik[58]   -308.8   -291.9   -250.2   -283.9   23.2  2.06        3       13
#> log_lik[59]    -32.7    -32.0    -28.8    -31.3    1.7  1.71        4       13
#> log_lik[60]   -179.0   -168.7   -144.3   -164.2   13.7  2.06        3       13
#> log_lik[61]  -2984.9  -2825.5  -2486.4  -2770.5  196.1  2.06        3       13
#> log_lik[62]     -7.4     -6.9     -6.4     -6.9    0.4  1.71        4       13
#> log_lik[63]   -753.7   -714.9   -626.5   -699.6   50.2  2.06        3       13
#> log_lik[64]    -37.6    -34.6    -30.4    -34.3    2.8  2.06        3       13
#> log_lik[65]   -651.7   -616.6   -545.1   -605.6   41.8  2.06        3       13
#> log_lik[66]  -1291.7  -1210.2  -1045.2  -1185.3   98.1  2.06        3       13
#> log_lik[67]    -44.5    -43.1    -39.0    -42.3    2.2  2.06        3       13
#> log_lik[68]   -676.1   -630.9   -542.7   -618.1   52.8  2.06        3       13
#> log_lik[69]   -119.7   -114.0   -104.0   -112.8    6.1  2.06        3       13
#> log_lik[70]  -5599.2  -5269.9  -4560.0  -5154.5  412.7  2.06        3       13
#> log_lik[71]    -57.1    -52.5    -43.8    -51.2    5.2  2.06        3       13
#> log_lik[72]  -2589.2  -2429.6  -2093.6  -2376.2  196.5  2.06        3       13
#> log_lik[73]   -997.1   -948.0   -821.3   -923.8   70.5  2.06        3       13
#> log_lik[74]  -2138.4  -1999.2  -1700.3  -1951.0  174.6  2.06        3       13
#> log_lik[75]   -593.1   -559.3   -479.7   -545.1   45.2  2.06        3       13
#> log_lik[76]   -797.5   -744.0   -629.7   -725.7   66.9  2.06        3       13
#> log_lik[77]    -75.4    -71.1    -66.6    -71.1    3.3  2.06        3       13
#> log_lik[78]    -13.2    -10.5     -6.3    -10.1    2.8  2.06        3       13
#> log_lik[79]    -14.1    -13.9    -13.8    -13.9    0.1  1.02       12       13
#> log_lik[80]    -13.5    -11.3     -8.0    -11.0    2.2  2.06        3       13
#> xstar[1,1]      -0.9      0.4      2.7      0.5    1.5  1.71        4       13
#> xstar[2,1]      -1.5      0.7      1.8      0.2    1.2  1.03       13       13
#> sigma[1]         2.0      2.1      2.2      2.1    0.1  2.06        3       13
#> lp__        -85033.1 -81240.8 -72606.6 -79753.4 4935.7  2.06        3       13
#> 
#> For each parameter, Bulk_ESS and Tail_ESS are crude measures of 
#> effective sample size for bulk and tail quantities respectively (an ESS > 100 
#> per chain is considered good), and Rhat is the potential scale reduction 
#> factor on rank normalized split chains (at convergence, Rhat <= 1.05).
fitted <- dfa_fitted(m)
# }