Get the fitted values from a DFA as a data frame

Get the fitted values from a DFA as a data frame

dfa_fitted(modelfit, conf_level = 0.95, names = NULL)

Arguments

modelfit

Output from fit_dfa.

conf_level

Probability level for CI.

names

Optional vector of names for time series labels. Should be same length as the number of time series.

Value

A data frame with the following columns: ID is an identifier for each time series, time is the time step, y is the observed values standardized to mean 0 and unit variance, estimate is the mean fitted value, lower is the lower CI, and upper is the upper CI.

See also

Examples

# \donttest{
y <- sim_dfa(num_trends = 2, num_years = 20, num_ts = 4)
m <- fit_dfa(y = y$y_sim, num_trends = 2, iter = 50, chains = 1)
#> 
#> SAMPLING FOR MODEL 'dfa' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000495 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 4.95 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: WARNING: There aren't enough warmup iterations to fit the
#> Chain 1:          three stages of adaptation as currently configured.
#> Chain 1:          Reducing each adaptation stage to 15%/75%/10% of
#> Chain 1:          the given number of warmup iterations:
#> Chain 1:            init_buffer = 3
#> Chain 1:            adapt_window = 20
#> Chain 1:            term_buffer = 2
#> Chain 1: 
#> Chain 1: Iteration:  1 / 50 [  2%]  (Warmup)
#> Chain 1: Iteration:  5 / 50 [ 10%]  (Warmup)
#> Chain 1: Iteration: 10 / 50 [ 20%]  (Warmup)
#> Chain 1: Iteration: 15 / 50 [ 30%]  (Warmup)
#> Chain 1: Iteration: 20 / 50 [ 40%]  (Warmup)
#> Chain 1: Iteration: 25 / 50 [ 50%]  (Warmup)
#> Chain 1: Iteration: 26 / 50 [ 52%]  (Sampling)
#> Chain 1: Iteration: 30 / 50 [ 60%]  (Sampling)
#> Chain 1: Iteration: 35 / 50 [ 70%]  (Sampling)
#> Chain 1: Iteration: 40 / 50 [ 80%]  (Sampling)
#> Chain 1: Iteration: 45 / 50 [ 90%]  (Sampling)
#> Chain 1: Iteration: 50 / 50 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.067791 seconds (Warm-up)
#> Chain 1:                0.106979 seconds (Sampling)
#> Chain 1:                0.17477 seconds (Total)
#> Chain 1: 
#> Warning: There were 1 chains where the estimated Bayesian Fraction of Missing Information was low. See
#> https://mc-stan.org/misc/warnings.html#bfmi-low
#> Warning: Examine the pairs() plot to diagnose sampling problems
#> Warning: The largest R-hat is NA, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#r-hat
#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#bulk-ess
#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#tail-ess
#> Inference for the input samples (1 chains: each with iter = 25; warmup = 12):
#> 
#>                    Q5       Q50      Q95      Mean      SD  Rhat Bulk_ESS
#> x[1,1]           -1.0      -1.0     -1.0      -1.0     0.0  2.06        3
#> x[2,1]            0.3       0.3      0.3       0.3     0.0  1.18        7
#> x[1,2]           -1.5      -1.5     -1.5      -1.5     0.0  2.06        3
#> x[2,2]           -1.1      -1.1     -1.1      -1.1     0.0  1.87        4
#> x[1,3]            0.4       0.5      0.5       0.5     0.0  2.06        3
#> x[2,3]            0.1       0.1      0.1       0.1     0.0  1.87        4
#> x[1,4]           -0.5      -0.5     -0.4      -0.5     0.0  2.06        3
#> x[2,4]            1.4       1.4      1.4       1.4     0.0  2.06        4
#> x[1,5]           -0.8      -0.8     -0.7      -0.8     0.0  2.06        3
#> x[2,5]            0.5       0.5      0.5       0.5     0.0  2.06        3
#> x[1,6]           -1.7      -1.7     -1.7      -1.7     0.0  2.06        3
#> x[2,6]           -0.2      -0.2     -0.2      -0.2     0.0  2.06        3
#> x[1,7]            0.1       0.1      0.1       0.1     0.0  2.06        3
#> x[2,7]           -0.9      -0.9     -0.8      -0.8     0.0  2.06        3
#> x[1,8]            1.9       1.9      1.9       1.9     0.0  2.06        4
#> x[2,8]           -2.0      -2.0     -1.9      -2.0     0.0  2.06        3
#> x[1,9]            0.9       0.9      0.9       0.9     0.0  2.06        3
#> x[2,9]           -1.2      -1.2     -1.1      -1.2     0.0  2.06        3
#> x[1,10]          -0.9      -0.9     -0.8      -0.9     0.0  2.06        3
#> x[2,10]          -0.6      -0.6     -0.6      -0.6     0.0  2.06        3
#> x[1,11]          -2.3      -2.3     -2.3      -2.3     0.0  2.06        3
#> x[2,11]          -1.4      -1.4     -1.4      -1.4     0.0  2.06        3
#> x[1,12]          -0.6      -0.6     -0.6      -0.6     0.0  2.06        3
#> x[2,12]           0.2       0.2      0.2       0.2     0.0  1.87        4
#> x[1,13]           0.3       0.3      0.3       0.3     0.0  1.19       11
#> x[2,13]          -0.2      -0.2     -0.2      -0.2     0.0  0.94       13
#> x[1,14]          -0.3      -0.3     -0.2      -0.3     0.0  1.18        7
#> x[2,14]           0.8       0.9      0.9       0.9     0.0  2.06        4
#> x[1,15]          -0.2      -0.2     -0.2      -0.2     0.0  2.06        3
#> x[2,15]           0.8       0.8      0.8       0.8     0.0  2.06        4
#> x[1,16]           1.4       1.4      1.4       1.4     0.0  2.06        3
#> x[2,16]           0.0       0.0      0.0       0.0     0.0  2.06        4
#> x[1,17]           0.5       0.5      0.5       0.5     0.0  2.06        3
#> x[2,17]          -0.1      -0.1     -0.1      -0.1     0.0  2.06        4
#> x[1,18]          -0.2      -0.1     -0.1      -0.2     0.0  2.06        3
#> x[2,18]           0.7       0.7      0.7       0.7     0.0  2.06        3
#> x[1,19]           0.7       0.8      0.8       0.7     0.0  2.06        3
#> x[2,19]           0.5       0.6      0.6       0.6     0.0  2.06        3
#> x[1,20]           0.3       0.4      0.4       0.4     0.0  2.06        3
#> x[2,20]           0.8       0.8      0.8       0.8     0.0  2.06        3
#> Z[1,1]          -92.3     -92.3    -92.2     -92.3     0.0  2.06        4
#> Z[2,1]           29.7      30.5     30.7      30.4     0.5  2.06        3
#> Z[3,1]           43.1      43.9     44.2      43.8     0.5  2.06        3
#> Z[4,1]          -62.8     -62.6    -61.9     -62.5     0.4  2.06        3
#> Z[1,2]            0.0       0.0      0.0       0.0     0.0  1.00       13
#> Z[2,2]          -99.7     -99.7    -99.7     -99.7     0.0  1.71        4
#> Z[3,2]          -22.0     -21.9    -21.5     -21.9     0.2  2.06        3
#> Z[4,2]            4.4       4.5      4.5       4.5     0.1  1.45        5
#> log_lik[1]    -2371.1   -2068.0  -1338.7   -1970.3   421.1  2.06        3
#> log_lik[2]     -848.9    -740.3   -476.8    -705.0   152.2  2.06        3
#> log_lik[3]     -702.2    -607.0   -386.0    -578.3   128.2  2.06        3
#> log_lik[4]    -1076.1    -933.2   -595.9    -888.5   194.9  2.06        3
#> log_lik[5]    -5197.0   -4524.7  -2924.6   -4311.3   925.8  2.06        3
#> log_lik[6]    -1129.6   -1010.6   -706.0    -968.5   176.5  2.06        3
#> log_lik[7]     -482.6    -413.9   -257.1    -393.3    90.8  2.06        3
#> log_lik[8]    -2122.8   -1835.2  -1170.7   -1748.2   385.6  2.06        3
#> log_lik[9]     -454.3    -419.1   -318.7    -404.8    58.9  2.06        3
#> log_lik[10]      -1.8      -1.5     -1.5      -1.6     0.2  2.06        4
#> log_lik[11]     -65.3     -60.1    -46.0     -58.1     8.3  2.06        3
#> log_lik[12]    -177.4    -163.3   -123.8    -157.7    23.2  2.06        3
#> log_lik[13]    -522.5    -437.6   -252.3    -414.0   106.8  2.06        3
#> log_lik[14]   -6565.5   -5775.5  -3810.9   -5508.1  1135.0  2.06        3
#> log_lik[15]    -721.8    -617.4   -382.1    -587.6   136.7  2.06        3
#> log_lik[16]    -367.9    -310.2   -181.4    -293.6    74.0  2.06        3
#> log_lik[17]   -1472.6   -1254.7   -764.4   -1191.8   284.1  2.06        3
#> log_lik[18]   -1436.9   -1267.3   -831.6   -1206.8   250.7  2.06        3
#> log_lik[19]    -590.5    -504.8   -311.0    -480.1   112.5  2.06        3
#> log_lik[20]    -730.5    -620.5   -375.2    -589.2   142.1  2.06        3
#> log_lik[21]   -6632.2   -5762.1  -3712.3   -5493.1  1189.6  2.06        3
#> log_lik[22]    -187.9    -167.9   -113.0    -160.1    32.0  2.06        3
#> log_lik[23]   -1339.9   -1157.4   -736.1   -1103.0   245.4  2.06        3
#> log_lik[24]   -2947.9   -2547.2  -1622.2   -2427.3   537.7  2.06        3
#> log_lik[25]     -10.0      -7.7     -6.9      -8.0     1.3  2.06        3
#> log_lik[26]   -2037.1   -1755.1  -1113.6   -1672.3   371.5  2.06        3
#> log_lik[27]    -118.6    -103.7    -70.1     -99.5    19.6  2.06        3
#> log_lik[28]     -19.0     -18.3    -15.9     -17.9     1.3  1.58        4
#> log_lik[29]   -7651.7   -6766.0  -4552.5   -6462.0  1276.9  2.06        3
#> log_lik[30]  -16734.1  -14569.2  -9411.8  -13882.4  2982.8  2.06        3
#> log_lik[31]   -4167.1   -3623.0  -2349.3   -3457.1   739.7  2.06        3
#> log_lik[32]   -4186.2   -3678.5  -2430.3   -3508.6   719.7  2.06        3
#> log_lik[33]   -1717.0   -1522.6  -1037.3   -1456.3   280.2  2.06        3
#> log_lik[34]   -5528.6   -4773.3  -3047.3   -4549.7  1003.4  2.06        3
#> log_lik[35]   -1079.3    -934.9   -605.4    -892.9   192.0  2.06        3
#> log_lik[36]    -934.3    -822.3   -548.2    -785.2   158.2  2.06        3
#> log_lik[37]   -1629.9   -1421.4   -911.3   -1352.3   294.6  2.06        3
#> log_lik[38]    -308.3    -260.0   -161.4    -248.2    58.2  2.06        3
#> log_lik[39]    -151.8    -133.1    -85.4    -126.4    27.4  2.06        3
#> log_lik[40]    -681.4    -591.4   -376.0    -562.6   124.7  2.06        3
#> log_lik[41]  -11781.3  -10331.2  -6773.3   -9850.5  2057.9  2.06        3
#> log_lik[42]   -1418.6   -1245.1   -842.0   -1192.6   236.3  2.06        3
#> log_lik[43]   -1326.4   -1155.4   -740.5   -1099.2   239.6  2.06        3
#> log_lik[44]   -5038.9   -4392.5  -2846.9   -4186.7   896.1  2.06        3
#> log_lik[45]    -779.0    -680.0   -430.5    -645.5   143.5  2.06        3
#> log_lik[46]    -402.5    -358.5   -238.5    -341.6    68.8  2.06        3
#> log_lik[47]    -240.5    -209.2   -132.2    -198.8    44.4  2.06        3
#> log_lik[48]    -398.8    -346.8   -218.0    -329.1    74.1  2.06        3
#> log_lik[49]    -188.4    -166.4   -117.5    -160.3    28.6  2.06        3
#> log_lik[50]    -181.6    -158.5   -108.5    -152.2    29.4  2.06        3
#> log_lik[51]     -66.5     -58.0    -40.2     -55.9    10.5  2.06        3
#> log_lik[52]     -81.7     -71.8    -50.3     -69.2    12.6  2.06        3
#> log_lik[53]    -135.4    -121.8    -81.7    -115.9    23.0  2.06        3
#> log_lik[54]   -2253.3   -1976.3  -1281.7   -1881.8   399.3  2.06        3
#> log_lik[55]    -245.8    -215.0   -138.7    -204.9    44.1  2.06        3
#> log_lik[56]    -112.0     -99.8    -65.7     -94.9    19.5  2.06        3
#> log_lik[57]    -106.2    -100.2    -73.8     -95.9    15.2  2.06        3
#> log_lik[58]   -1877.6   -1646.3  -1061.9   -1567.0   335.9  2.06        3
#> log_lik[59]    -184.3    -163.6   -108.2    -156.1    31.9  2.06        3
#> log_lik[60]     -77.2     -71.5    -50.6     -68.2    11.9  2.06        3
#> log_lik[61]   -4219.8   -3656.1  -2365.2   -3489.3   750.5  2.06        3
#> log_lik[62]    -471.5    -416.6   -285.0    -398.5    76.2  2.06        3
#> log_lik[63]    -990.7    -853.8   -547.5    -814.8   178.6  2.06        3
#> log_lik[64]   -1944.6   -1676.1  -1072.4   -1599.0   351.5  2.06        3
#> log_lik[65]    -580.0    -482.6   -288.2    -459.7   113.8  2.06        3
#> log_lik[66]    -112.4    -102.8    -78.1     -99.0    14.2  2.06        3
#> log_lik[67]    -150.0    -126.2    -78.2    -120.5    28.1  2.06        3
#> log_lik[68]    -283.2    -235.4   -140.7    -224.3    55.5  2.06        3
#> log_lik[69]     -40.3     -37.5    -34.4     -37.7     1.9  1.32        5
#> log_lik[70]   -1562.8   -1349.0   -851.7   -1286.4   288.9  2.06        3
#> log_lik[71]    -106.6     -96.4    -67.6     -92.5    16.8  2.06        3
#> log_lik[72]     -30.4     -29.8    -26.3     -29.2     2.1  0.94       10
#> log_lik[73]   -1362.7   -1148.1   -704.2   -1095.2   259.9  2.06        3
#> log_lik[74]    -343.0    -289.9   -174.4    -276.4    67.4  2.06        3
#> log_lik[75]    -133.5    -112.2    -69.3    -107.0    25.2  2.06        3
#> log_lik[76]    -542.4    -453.8   -275.8    -433.0   104.6  2.06        3
#> log_lik[77]    -360.9    -290.4   -162.6    -276.4    75.4  2.06        3
#> log_lik[78]   -1347.4   -1158.6   -724.3   -1105.1   252.1  2.06        3
#> log_lik[79]      -1.6      -1.4     -1.3      -1.4     0.2  1.01        7
#> log_lik[80]    -134.0    -106.8    -59.6    -101.8    28.1  2.06        3
#> xstar[1,1]       -1.1       1.1      2.7       1.1     1.3  1.27        5
#> xstar[2,1]       -0.5       1.0      2.4       1.1     1.0  1.33        6
#> sigma[1]          1.4       1.5      1.9       1.6     0.3  2.06        3
#> lp__        -139602.2 -123238.3 -84328.5 -118084.7 22535.3  2.06        3
#>             Tail_ESS
#> x[1,1]            13
#> x[2,1]            13
#> x[1,2]            13
#> x[2,2]            13
#> x[1,3]            13
#> x[2,3]            13
#> x[1,4]            13
#> x[2,4]            13
#> x[1,5]            13
#> x[2,5]            13
#> x[1,6]            13
#> x[2,6]            13
#> x[1,7]            13
#> x[2,7]            13
#> x[1,8]            13
#> x[2,8]            13
#> x[1,9]            13
#> x[2,9]            13
#> x[1,10]           13
#> x[2,10]           13
#> x[1,11]           13
#> x[2,11]           13
#> x[1,12]           13
#> x[2,12]           13
#> x[1,13]           13
#> x[2,13]           13
#> x[1,14]           13
#> x[2,14]           13
#> x[1,15]           13
#> x[2,15]           13
#> x[1,16]           13
#> x[2,16]           13
#> x[1,17]           13
#> x[2,17]           13
#> x[1,18]           13
#> x[2,18]           13
#> x[1,19]           13
#> x[2,19]           13
#> x[1,20]           13
#> x[2,20]           13
#> Z[1,1]            13
#> Z[2,1]            13
#> Z[3,1]            13
#> Z[4,1]            13
#> Z[1,2]            13
#> Z[2,2]            13
#> Z[3,2]            13
#> Z[4,2]            13
#> log_lik[1]        13
#> log_lik[2]        13
#> log_lik[3]        13
#> log_lik[4]        13
#> log_lik[5]        13
#> log_lik[6]        13
#> log_lik[7]        13
#> log_lik[8]        13
#> log_lik[9]        13
#> log_lik[10]       13
#> log_lik[11]       13
#> log_lik[12]       13
#> log_lik[13]       13
#> log_lik[14]       13
#> log_lik[15]       13
#> log_lik[16]       13
#> log_lik[17]       13
#> log_lik[18]       13
#> log_lik[19]       13
#> log_lik[20]       13
#> log_lik[21]       13
#> log_lik[22]       13
#> log_lik[23]       13
#> log_lik[24]       13
#> log_lik[25]       13
#> log_lik[26]       13
#> log_lik[27]       13
#> log_lik[28]       13
#> log_lik[29]       13
#> log_lik[30]       13
#> log_lik[31]       13
#> log_lik[32]       13
#> log_lik[33]       13
#> log_lik[34]       13
#> log_lik[35]       13
#> log_lik[36]       13
#> log_lik[37]       13
#> log_lik[38]       13
#> log_lik[39]       13
#> log_lik[40]       13
#> log_lik[41]       13
#> log_lik[42]       13
#> log_lik[43]       13
#> log_lik[44]       13
#> log_lik[45]       13
#> log_lik[46]       13
#> log_lik[47]       13
#> log_lik[48]       13
#> log_lik[49]       13
#> log_lik[50]       13
#> log_lik[51]       13
#> log_lik[52]       13
#> log_lik[53]       13
#> log_lik[54]       13
#> log_lik[55]       13
#> log_lik[56]       13
#> log_lik[57]       13
#> log_lik[58]       13
#> log_lik[59]       13
#> log_lik[60]       13
#> log_lik[61]       13
#> log_lik[62]       13
#> log_lik[63]       13
#> log_lik[64]       13
#> log_lik[65]       13
#> log_lik[66]       13
#> log_lik[67]       13
#> log_lik[68]       13
#> log_lik[69]       13
#> log_lik[70]       13
#> log_lik[71]       13
#> log_lik[72]       13
#> log_lik[73]       13
#> log_lik[74]       13
#> log_lik[75]       13
#> log_lik[76]       13
#> log_lik[77]       13
#> log_lik[78]       13
#> log_lik[79]       13
#> log_lik[80]       13
#> xstar[1,1]        13
#> xstar[2,1]        13
#> sigma[1]          13
#> lp__              13
#> 
#> For each parameter, Bulk_ESS and Tail_ESS are crude measures of 
#> effective sample size for bulk and tail quantities respectively (an ESS > 100 
#> per chain is considered good), and Rhat is the potential scale reduction 
#> factor on rank normalized split chains (at convergence, Rhat <= 1.05).
fitted <- dfa_fitted(m)
# }