Get the fitted values from a DFA as a data frame

Get the fitted values from a DFA as a data frame

dfa_fitted(modelfit, conf_level = 0.95, names = NULL)

Arguments

modelfit

Output from fit_dfa.

conf_level

Probability level for CI.

names

Optional vector of names for time series labels. Should be same length as the number of time series.

Value

A data frame with the following columns: ID is an identifier for each time series, time is the time step, y is the observed values standardized to mean 0 and unit variance, estimate is the mean fitted value, lower is the lower CI, and upper is the upper CI.

See also

Examples

# \donttest{
y <- sim_dfa(num_trends = 2, num_years = 20, num_ts = 4)
m <- fit_dfa(y = y$y_sim, num_trends = 2, iter = 50, chains = 1)
#> 
#> SAMPLING FOR MODEL 'dfa' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.00012 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.2 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: WARNING: There aren't enough warmup iterations to fit the
#> Chain 1:          three stages of adaptation as currently configured.
#> Chain 1:          Reducing each adaptation stage to 15%/75%/10% of
#> Chain 1:          the given number of warmup iterations:
#> Chain 1:            init_buffer = 3
#> Chain 1:            adapt_window = 20
#> Chain 1:            term_buffer = 2
#> Chain 1: 
#> Chain 1: Iteration:  1 / 50 [  2%]  (Warmup)
#> Chain 1: Iteration:  5 / 50 [ 10%]  (Warmup)
#> Chain 1: Iteration: 10 / 50 [ 20%]  (Warmup)
#> Chain 1: Iteration: 15 / 50 [ 30%]  (Warmup)
#> Chain 1: Iteration: 20 / 50 [ 40%]  (Warmup)
#> Chain 1: Iteration: 25 / 50 [ 50%]  (Warmup)
#> Chain 1: Iteration: 26 / 50 [ 52%]  (Sampling)
#> Chain 1: Iteration: 30 / 50 [ 60%]  (Sampling)
#> Chain 1: Iteration: 35 / 50 [ 70%]  (Sampling)
#> Chain 1: Iteration: 40 / 50 [ 80%]  (Sampling)
#> Chain 1: Iteration: 45 / 50 [ 90%]  (Sampling)
#> Chain 1: Iteration: 50 / 50 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.006072 seconds (Warm-up)
#> Chain 1:                0.00827 seconds (Sampling)
#> Chain 1:                0.014342 seconds (Total)
#> Chain 1: 
#> Warning: There were 1 chains where the estimated Bayesian Fraction of Missing Information was low. See
#> https://mc-stan.org/misc/warnings.html#bfmi-low
#> Warning: Examine the pairs() plot to diagnose sampling problems
#> Warning: The largest R-hat is NA, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#r-hat
#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#bulk-ess
#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#tail-ess
#> Inference for the input samples (1 chains: each with iter = 25; warmup = 12):
#> 
#>                   Q5      Q50      Q95     Mean     SD  Rhat Bulk_ESS Tail_ESS
#> x[1,1]          -1.9     -1.8     -1.8     -1.8    0.0  1.71        5       13
#> x[2,1]          -2.0     -2.0     -2.0     -2.0    0.0  1.87        5       13
#> x[1,2]          -0.2     -0.2     -0.2     -0.2    0.0  2.06        3       13
#> x[2,2]          -2.0     -1.9     -1.9     -1.9    0.0  1.87        5       13
#> x[1,3]           1.6      1.7      1.7      1.7    0.0  1.71        4       13
#> x[2,3]          -1.5     -1.5     -1.5     -1.5    0.0  1.71        4       13
#> x[1,4]           1.1      1.1      1.1      1.1    0.0  2.06        3       13
#> x[2,4]          -3.0     -3.0     -3.0     -3.0    0.0  2.06        3       13
#> x[1,5]           1.5      1.5      1.6      1.5    0.0  2.06        3       13
#> x[2,5]          -2.3     -2.3     -2.3     -2.3    0.0  2.06        3       13
#> x[1,6]           1.0      1.0      1.1      1.0    0.0  2.06        3       13
#> x[2,6]          -3.1     -3.0     -3.0     -3.0    0.0  2.06        3       13
#> x[1,7]           2.0      2.0      2.0      2.0    0.0  2.06        3       13
#> x[2,7]          -1.8     -1.8     -1.8     -1.8    0.0  2.06        3       13
#> x[1,8]           3.8      3.9      3.9      3.9    0.0  2.06        3       13
#> x[2,8]          -2.0     -1.9     -1.9     -1.9    0.0  2.06        3       13
#> x[1,9]           4.5      4.5      4.6      4.5    0.0  1.87        4       13
#> x[2,9]          -1.1     -1.0     -0.9     -1.0    0.0  2.06        3       13
#> x[1,10]          2.7      2.7      2.7      2.7    0.0  1.87        4       13
#> x[2,10]         -0.1      0.0      0.1      0.0    0.1  2.06        3       13
#> x[1,11]          2.3      2.4      2.4      2.4    0.0  1.87        4       13
#> x[2,11]          1.0      1.2      1.2      1.1    0.1  2.06        3       13
#> x[1,12]          0.7      0.8      0.8      0.8    0.0  1.58        4       13
#> x[2,12]          2.1      2.3      2.3      2.3    0.1  2.06        3       13
#> x[1,13]          1.5      1.5      1.5      1.5    0.0  1.58        4       13
#> x[2,13]          1.9      2.0      2.1      2.0    0.1  2.06        3       13
#> x[1,14]          1.6      1.6      1.7      1.6    0.0  1.71        4       13
#> x[2,14]         -0.4     -0.2     -0.2     -0.3    0.1  2.06        3       13
#> x[1,15]         -0.2     -0.1     -0.1     -0.1    0.0  1.87        5       13
#> x[2,15]          1.1      1.3      1.4      1.3    0.1  2.06        3       13
#> x[1,16]         -1.5     -1.5     -1.5     -1.5    0.0  2.06        6       13
#> x[2,16]          2.1      2.4      2.4      2.3    0.1  2.06        3       13
#> x[1,17]         -3.3     -3.3     -3.3     -3.3    0.0  2.06        6       13
#> x[2,17]          3.7      4.0      4.0      3.9    0.1  2.06        3       13
#> x[1,18]         -1.2     -1.2     -1.2     -1.2    0.0  1.87        6       13
#> x[2,18]          4.5      4.8      4.8      4.7    0.1  2.06        3       13
#> x[1,19]         -0.5     -0.5     -0.5     -0.5    0.0  1.05       13       13
#> x[2,19]          5.2      5.5      5.6      5.4    0.1  2.06        3       13
#> x[1,20]          1.0      1.0      1.0      1.0    0.0  0.95       13       13
#> x[2,20]          4.0      4.3      4.4      4.2    0.1  2.06        3       13
#> Z[1,1]         -94.2    -94.2    -94.2    -94.2    0.0  1.58        4       13
#> Z[2,1]          52.1     53.0     53.2     52.8    0.4  2.06        3       13
#> Z[3,1]         -56.7    -56.6    -56.0    -56.4    0.3  2.06        3       13
#> Z[4,1]          -6.9     -6.7     -6.2     -6.6    0.3  2.06        3       13
#> Z[1,2]           0.0      0.0      0.0      0.0    0.0  1.00       13       13
#> Z[2,2]         -98.6    -98.6    -98.6    -98.6    0.0  1.71        4       13
#> Z[3,2]         -28.6    -28.5    -27.9    -28.3    0.3  2.06        3       13
#> Z[4,2]          66.5     67.1     67.3     66.9    0.3  2.06        3       13
#> log_lik[1]    -240.7   -217.6   -151.1   -201.6   34.8  2.06        3       13
#> log_lik[2]     -80.0    -72.7    -51.8    -67.7   10.9  2.06        3       13
#> log_lik[3]    -209.4   -188.0   -126.6   -173.2   32.2  2.06        3       13
#> log_lik[4]    -126.2   -113.7    -78.6   -105.5   18.4  2.06        3       13
#> log_lik[5]      -5.0     -5.0     -4.9     -5.0    0.0  0.96       13       13
#> log_lik[6]    -267.7   -240.8   -163.9   -222.5   40.4  2.06        3       13
#> log_lik[7]     -36.6    -33.4    -24.1    -31.2    4.9  2.06        3       13
#> log_lik[8]    -143.2   -128.5    -87.6   -118.9   21.5  2.06        3       13
#> log_lik[9]    -209.8   -187.5   -122.9   -171.9   33.9  2.06        3       13
#> log_lik[10]   -445.0   -399.9   -268.4   -368.3   68.8  2.06        3       13
#> log_lik[11]    -27.1    -24.3    -16.3    -22.4    4.2  2.06        3       13
#> log_lik[12]   -104.1    -93.5    -63.6    -86.4   15.7  2.06        3       13
#> log_lik[13]    -99.4    -88.1    -55.9    -80.3   16.9  2.06        3       13
#> log_lik[14]  -1014.3   -912.3   -615.5   -841.2  155.2  2.06        3       13
#> log_lik[15]     -5.9     -5.8     -5.6     -5.7    0.1  2.06        4       13
#> log_lik[16]   -354.0   -317.7   -214.5   -293.2   54.2  2.06        3       13
#> log_lik[17]   -178.5   -157.8   -100.1   -144.0   30.5  2.06        3       13
#> log_lik[18]   -771.5   -692.8   -466.7   -638.9  118.4  2.06        3       13
#> log_lik[19]     -7.4     -6.7     -5.0     -6.3    1.0  2.06        3       13
#> log_lik[20]   -220.9   -198.3   -134.9   -183.4   33.3  2.06        3       13
#> log_lik[21]    -85.2    -74.5    -45.9    -67.7   15.3  2.06        3       13
#> log_lik[22]  -1023.8   -921.6   -626.1   -851.0  154.7  2.06        3       13
#> log_lik[23]     -8.5     -8.4     -7.9     -8.3    0.2  2.06        4       13
#> log_lik[24]   -362.7   -326.4   -222.8   -301.8   54.2  2.06        3       13
#> log_lik[25]   -301.9   -267.8   -170.2   -244.4   51.3  2.06        3       13
#> log_lik[26]   -657.7   -592.8   -404.5   -548.0   98.5  2.06        3       13
#> log_lik[27]    -36.9    -32.4    -19.7    -29.3    6.7  2.06        3       13
#> log_lik[28]   -147.6   -133.5    -93.6   -124.0   21.0  2.06        3       13
#> log_lik[29]  -1093.8   -976.9   -636.2   -894.7  178.2  2.06        3       13
#> log_lik[30]  -1261.1  -1136.4   -771.0  -1049.1  190.7  2.06        3       13
#> log_lik[31]   -228.1   -202.1   -125.8   -183.4   40.0  2.06        3       13
#> log_lik[32]   -194.3   -175.4   -121.9   -162.7   28.1  2.06        3       13
#> log_lik[33]  -1504.9  -1345.4   -882.7  -1234.0  242.2  2.06        3       13
#> log_lik[34]   -924.1   -834.5   -572.2   -772.0  136.9  2.06        3       13
#> log_lik[35]   -438.8   -389.2   -244.7   -354.0   75.7  2.06        3       13
#> log_lik[36]    -77.0    -70.3    -51.5    -65.9    9.9  2.06        3       13
#> log_lik[37]   -543.1   -483.5   -312.5   -442.5   89.7  2.06        3       13
#> log_lik[38]   -159.7   -146.6   -108.1   -137.7   20.1  2.06        3       13
#> log_lik[39]   -203.7   -179.3   -109.7   -162.4   36.6  2.06        3       13
#> log_lik[40]     -5.4     -4.8     -4.6     -5.0    0.3  2.06        3       13
#> log_lik[41]   -410.0   -364.8   -235.3   -333.7   68.0  2.06        3       13
#> log_lik[42]     -5.5     -4.4     -4.1     -4.7    0.6  2.06        3       13
#> log_lik[43]   -232.3   -204.2   -124.4   -184.9   42.0  2.06        3       13
#> log_lik[44]    -36.2    -31.2    -17.8    -28.0    7.1  2.06        3       13
#> log_lik[45]    -47.9    -42.2    -26.1    -38.4    8.5  2.06        3       13
#> log_lik[46]   -286.7   -250.7   -149.5   -226.1   53.2  2.06        3       13
#> log_lik[47]   -103.3    -89.8    -52.3    -80.8   19.9  2.06        3       13
#> log_lik[48]   -190.5   -165.8    -97.6   -149.4   35.9  2.06        3       13
#> log_lik[49]   -168.0   -149.6    -96.6   -136.9   27.7  2.06        3       13
#> log_lik[50]   -128.0   -109.6    -59.4    -97.4   26.6  2.06        3       13
#> log_lik[51]   -174.5   -152.8    -91.3   -137.9   32.4  2.06        3       13
#> log_lik[52]   -141.7   -122.3    -69.5   -109.7   28.0  2.06        3       13
#> log_lik[53]   -200.7   -179.6   -118.2   -164.9   32.0  2.06        3       13
#> log_lik[54]    -91.0    -88.8    -80.6    -86.8    4.1  1.87        4       13
#> log_lik[55]    -67.8    -59.0    -34.3    -53.0   13.0  2.06        3       13
#> log_lik[56]    -10.1     -7.6     -7.0     -8.3    1.2  2.06        3       13
#> log_lik[57]     -4.2     -4.0     -3.9     -4.0    0.1  2.06        6       13
#> log_lik[58]   -168.3   -141.9    -71.0   -124.7   37.8  2.06        3       13
#> log_lik[59]    -11.9    -10.1     -5.7     -9.1    2.4  2.06        3       13
#> log_lik[60]    -75.1    -62.9    -31.0    -55.2   17.1  2.06        3       13
#> log_lik[61]   -171.7   -155.8   -109.2   -144.5   24.3  2.06        3       13
#> log_lik[62]   -833.5   -726.0   -420.2   -651.7  160.6  2.06        3       13
#> log_lik[63]     -7.1     -5.8     -5.5     -6.1    0.6  1.58        4       13
#> log_lik[64]   -248.0   -212.9   -116.4   -189.6   51.1  2.06        3       13
#> log_lik[65]   -772.7   -697.0   -475.8   -643.5  115.3  2.06        3       13
#> log_lik[66]  -2672.5  -2356.6  -1445.7  -2135.6  476.9  2.06        3       13
#> log_lik[67]    -42.0    -40.2    -34.5    -38.8    2.9  2.06        3       13
#> log_lik[68]   -715.1   -622.4   -364.0   -560.0  136.3  2.06        3       13
#> log_lik[69]   -113.2   -102.3    -71.1    -94.8   16.4  2.06        3       13
#> log_lik[70]  -2395.2  -2108.8  -1286.4  -1909.2  431.3  2.06        3       13
#> log_lik[71]    -41.0    -35.0    -18.8    -31.1    8.7  2.06        3       13
#> log_lik[72]   -916.9   -800.7   -475.5   -722.4  171.4  2.06        3       13
#> log_lik[73]    -21.3    -19.5    -14.5    -18.3    2.6  2.06        3       13
#> log_lik[74]  -2695.9  -2375.4  -1458.2  -2153.2  481.2  2.06        3       13
#> log_lik[75]   -141.7   -123.3    -71.3   -110.7   27.5  2.06        3       13
#> log_lik[76]  -1167.2  -1021.8   -614.9   -924.0  214.3  2.06        3       13
#> log_lik[77]    -80.0    -72.7    -50.2    -67.2   11.6  2.06        3       13
#> log_lik[78]  -1149.1  -1003.0   -585.8   -901.5  219.2  2.06        3       13
#> log_lik[79]   -277.1   -243.6   -146.7   -219.9   51.0  2.06        3       13
#> log_lik[80]   -671.4   -584.9   -342.0   -526.3  127.9  2.06        3       13
#> xstar[1,1]       0.0      1.2      2.2      1.2    0.7  1.07       13       13
#> xstar[2,1]       3.3      4.3      6.0      4.3    1.0  1.32       13       13
#> sigma[1]         7.8      8.3     10.0      8.7    0.9  2.06        3       13
#> lp__        -46513.6 -42931.0 -32635.2 -40448.0 5396.7  2.06        3       13
#> 
#> For each parameter, Bulk_ESS and Tail_ESS are crude measures of 
#> effective sample size for bulk and tail quantities respectively (an ESS > 100 
#> per chain is considered good), and Rhat is the potential scale reduction 
#> factor on rank normalized split chains (at convergence, Rhat <= 1.05).
fitted <- dfa_fitted(m)
# }