Get the fitted values from a DFA as a data frame

Get the fitted values from a DFA as a data frame

dfa_fitted(modelfit, conf_level = 0.95, names = NULL)

Arguments

modelfit
Output from fit_dfa.

conf_level
Probability level for CI.

names
Optional vector of names for time series labels. Should be same length as the number of time series.

Value

A data frame with the following columns: ID is an identifier for each time series, time is the time step, y is the observed values standardized to mean 0 and unit variance, estimate is the mean fitted value, lower is the lower CI, and upper is the upper CI.

See also

Examples

# \donttest{
y <- sim_dfa(num_trends = 2, num_years = 20, num_ts = 4)
m <- fit_dfa(y = y$y_sim, num_trends = 2, iter = 50, chains = 1)

#> 
#> SAMPLING FOR MODEL 'dfa' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000128 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.28 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: WARNING: There aren't enough warmup iterations to fit the
#> Chain 1:          three stages of adaptation as currently configured.
#> Chain 1:          Reducing each adaptation stage to 15%/75%/10% of
#> Chain 1:          the given number of warmup iterations:
#> Chain 1:            init_buffer = 3
#> Chain 1:            adapt_window = 20
#> Chain 1:            term_buffer = 2
#> Chain 1: 
#> Chain 1: Iteration:  1 / 50 [  2%]  (Warmup)
#> Chain 1: Iteration:  5 / 50 [ 10%]  (Warmup)
#> Chain 1: Iteration: 10 / 50 [ 20%]  (Warmup)
#> Chain 1: Iteration: 15 / 50 [ 30%]  (Warmup)
#> Chain 1: Iteration: 20 / 50 [ 40%]  (Warmup)
#> Chain 1: Iteration: 25 / 50 [ 50%]  (Warmup)
#> Chain 1: Iteration: 26 / 50 [ 52%]  (Sampling)
#> Chain 1: Iteration: 30 / 50 [ 60%]  (Sampling)
#> Chain 1: Iteration: 35 / 50 [ 70%]  (Sampling)
#> Chain 1: Iteration: 40 / 50 [ 80%]  (Sampling)
#> Chain 1: Iteration: 45 / 50 [ 90%]  (Sampling)
#> Chain 1: Iteration: 50 / 50 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.006084 seconds (Warm-up)
#> Chain 1:                0.012601 seconds (Sampling)
#> Chain 1:                0.018685 seconds (Total)
#> Chain 1: 

#> Warning: There were 1 chains where the estimated Bayesian Fraction of Missing Information was low. See
#> http://mc-stan.org/misc/warnings.html#bfmi-low

#> Warning: Examine the pairs() plot to diagnose sampling problems

#> Warning: The largest R-hat is NA, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#r-hat

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#tail-ess

#> Inference for the input samples (1 chains: each with iter = 25; warmup = 12):
#> 
#>                   Q5      Q50     Q95     Mean     SD  Rhat Bulk_ESS Tail_ESS
#> x[1,1]          -0.7     -0.7    -0.7     -0.7    0.0  2.06        4       13
#> x[2,1]          -0.3     -0.3    -0.2     -0.2    0.0  1.87        7       13
#> x[1,2]          -0.9     -0.9    -0.8     -0.9    0.0  2.06        4       13
#> x[2,2]           0.1      0.1     0.1      0.1    0.0  2.06        4       13
#> x[1,3]          -1.9     -1.8    -1.8     -1.9    0.0  2.06        4       13
#> x[2,3]           0.5      0.5     0.5      0.5    0.0  2.06        4       13
#> x[1,4]          -0.6     -0.5    -0.5     -0.5    0.0  2.06        4       13
#> x[2,4]           0.7      0.7     0.8      0.7    0.0  2.06        4       13
#> x[1,5]           0.4      0.5     0.5      0.5    0.0  2.06        4       13
#> x[2,5]          -0.2     -0.2    -0.1     -0.2    0.0  2.06        4       13
#> x[1,6]          -0.7     -0.6    -0.6     -0.6    0.1  2.06        4       13
#> x[2,6]          -1.4     -1.4    -1.3     -1.4    0.1  2.06        4       13
#> x[1,7]           0.6      0.8     0.8      0.7    0.1  2.06        4       13
#> x[2,7]          -1.7     -1.7    -1.5     -1.7    0.1  2.06        4       13
#> x[1,8]           0.2      0.3     0.3      0.3    0.1  2.06        4       13
#> x[2,8]          -2.0     -2.0    -1.9     -2.0    0.1  2.06        4       13
#> x[1,9]           0.4      0.6     0.6      0.5    0.1  2.06        4       13
#> x[2,9]          -2.6     -2.6    -2.4     -2.6    0.1  2.06        4       13
#> x[1,10]         -0.2      0.0     0.0      0.0    0.1  2.06        4       13
#> x[2,10]         -3.9     -3.9    -3.7     -3.9    0.1  2.06        4       13
#> x[1,11]         -0.6     -0.5    -0.5     -0.5    0.1  2.06        4       13
#> x[2,11]         -3.1     -3.1    -2.9     -3.1    0.1  2.06        4       13
#> x[1,12]         -1.8     -1.6    -1.6     -1.7    0.1  2.06        4       13
#> x[2,12]         -4.2     -4.2    -4.0     -4.2    0.1  2.06        4       13
#> x[1,13]         -2.6     -2.5    -2.5     -2.5    0.0  1.58        4       13
#> x[2,13]         -4.9     -4.8    -4.7     -4.8    0.1  2.06        4       13
#> x[1,14]         -3.0     -2.9    -2.9     -3.0    0.0  1.20        9       13
#> x[2,14]         -4.5     -4.5    -4.3     -4.5    0.1  2.06        4       13
#> x[1,15]         -4.3     -4.2    -4.2     -4.2    0.0  1.39       10       13
#> x[2,15]         -3.8     -3.8    -3.6     -3.7    0.1  2.06        4       13
#> x[1,16]         -5.3     -5.2    -5.2     -5.2    0.0  1.47       10       13
#> x[2,16]         -4.6     -4.6    -4.4     -4.5    0.1  2.06        4       13
#> x[1,17]         -5.9     -5.9    -5.9     -5.9    0.0  1.47        9       13
#> x[2,17]         -5.0     -5.0    -4.8     -4.9    0.1  2.06        4       13
#> x[1,18]         -7.0     -7.0    -7.0     -7.0    0.0  1.47        9       13
#> x[2,18]         -3.9     -3.9    -3.8     -3.9    0.1  2.06        4       13
#> x[1,19]         -7.3     -7.3    -7.3     -7.3    0.0  1.21       11       13
#> x[2,19]         -4.6     -4.6    -4.5     -4.6    0.0  2.06        3       13
#> x[1,20]         -6.6     -6.6    -6.5     -6.6    0.0  1.16       12       13
#> x[2,20]         -4.0     -4.0    -3.9     -4.0    0.0  2.06        3       13
#> Z[1,1]           4.4      4.5     4.5      4.4    0.0  1.15        8       13
#> Z[2,1]          -5.5     14.9    18.2     11.4    8.8  2.06        3       13
#> Z[3,1]          -9.7     -9.3    -4.2     -8.2    2.3  2.06        4       13
#> Z[4,1]         -50.8    -49.0   -26.4    -44.4   10.1  2.06        3       13
#> Z[1,2]           0.0      0.0     0.0      0.0    0.0  1.00       13       13
#> Z[2,2]           0.8      0.8     0.8      0.8    0.0  2.06       13       13
#> Z[3,2]           4.8      8.9     9.5      8.2    1.9  2.06        3       13
#> Z[4,2]          29.8     50.2    51.3     46.0    9.2  2.06        3       13
#> log_lik[1]      -2.8     -2.4    -2.4     -2.5    0.1  2.06        3       13
#> log_lik[2]      -8.5     -5.9    -2.9     -5.9    2.1  2.06        3       13
#> log_lik[3]      -2.9     -2.8    -2.7     -2.8    0.1  2.06        4       13
#> log_lik[4]     -27.5    -23.0    -5.8    -20.1    8.4  2.06        3       13
#> log_lik[5]      -2.8     -2.4    -2.4     -2.5    0.1  2.06        3       13
#> log_lik[6]      -9.9     -6.7    -3.0     -6.6    2.6  2.06        3       13
#> log_lik[7]      -5.6     -5.1    -3.0     -4.8    1.0  2.06        3       13
#> log_lik[8]     -96.5    -81.8   -16.1    -70.3   31.2  2.06        3       13
#> log_lik[9]      -3.9     -3.8    -3.4     -3.7    0.2  2.06        3       13
#> log_lik[10]    -41.5    -25.0    -4.0    -24.2   13.9  2.06        3       13
#> log_lik[11]    -22.1    -18.6    -4.6    -16.2    6.8  2.06        3       13
#> log_lik[12]   -538.7   -454.5   -74.9   -387.8  180.4  2.06        3       13
#> log_lik[13]     -2.7     -2.3    -2.3     -2.4    0.2  2.06        3       13
#> log_lik[14]     -4.8     -3.7    -2.9     -3.8    0.7  2.06        3       13
#> log_lik[15]     -7.8     -6.8    -3.3     -6.3    1.7  2.06        3       13
#> log_lik[16]   -162.6   -139.6   -29.6   -120.1   51.9  2.06        3       13
#> log_lik[17]     -2.8     -2.7    -2.7     -2.7    0.1  1.58        9       13
#> log_lik[18]     -6.4     -4.9    -2.6     -4.7    1.4  2.06        3       13
#> log_lik[19]     -4.2     -3.9    -2.8     -3.7    0.6  2.06        4       13
#> log_lik[20]    -49.7    -41.3    -5.3    -34.9   17.3  2.06        3       13
#> log_lik[21]     -2.8     -2.5    -2.5     -2.6    0.1  2.06        3       13
#> log_lik[22]     -7.8     -5.7    -2.8     -5.6    1.9  2.06        3       13
#> log_lik[23]     -4.8     -4.3    -2.9     -4.1    0.7  2.06        3       13
#> log_lik[24]    -68.9    -59.7    -9.0    -50.0   23.6  2.06        3       13
#> log_lik[25]     -2.9     -2.9    -2.9     -2.9    0.0  1.33        5       13
#> log_lik[26]     -9.2     -6.3    -2.8     -6.2    2.4  2.06        3       13
#> log_lik[27]    -22.8    -18.8    -4.2    -16.4    7.2  2.06        3       13
#> log_lik[28]   -608.1   -514.1   -66.9   -433.3  211.4  2.06        3       13
#> log_lik[29]     -2.7     -2.5    -2.5     -2.5    0.1  2.06        3       13
#> log_lik[30]     -3.4     -2.9    -2.7     -3.0    0.3  2.06        3       13
#> log_lik[31]    -22.4    -18.4    -4.4    -16.1    6.9  2.06        3       13
#> log_lik[32]   -562.7   -477.6   -62.3   -401.8  195.7  2.06        3       13
#> log_lik[33]     -2.8     -2.6    -2.6     -2.7    0.1  1.10        8       13
#> log_lik[34]     -5.9     -4.4    -2.8     -4.4    1.2  2.06        3       13
#> log_lik[35]    -42.1    -34.4    -6.5    -29.9   13.7  2.06        3       13
#> log_lik[36]  -1027.8   -870.7  -118.0   -734.5  355.5  2.06        3       13
#> log_lik[37]     -2.7     -2.3    -2.2     -2.3    0.2  2.06        3       13
#> log_lik[38]     -2.7     -2.5    -2.5     -2.5    0.1  2.06        3       13
#> log_lik[39]    -61.8    -50.0    -9.0    -43.5   20.3  2.06        3       13
#> log_lik[40]  -1545.9  -1318.0  -190.9  -1112.6  530.3  2.06        3       13
#> log_lik[41]     -2.8     -2.5    -2.5     -2.6    0.2  2.06        3       13
#> log_lik[42]     -6.9     -5.4    -2.6     -5.2    1.6  2.06        3       13
#> log_lik[43]    -27.6    -22.3    -5.2    -19.7    8.6  2.06        3       13
#> log_lik[44]   -727.3   -621.5   -89.3   -523.6  249.9  2.06        3       13
#> log_lik[45]     -4.5     -4.4    -3.7     -4.3    0.3  2.06        3       13
#> log_lik[46]    -45.1    -29.0    -3.1    -26.9   15.6  2.06        3       13
#> log_lik[47]    -24.3    -19.5    -5.1    -17.4    7.3  2.06        3       13
#> log_lik[48]   -702.2   -609.4   -96.2   -512.9  238.7  2.06        3       13
#> log_lik[49]     -7.9     -7.4    -5.0     -7.0    1.1  2.06        3       13
#> log_lik[50]   -100.3    -63.1    -3.6    -58.4   35.9  2.06        3       13
#> log_lik[51]    -17.7    -14.1    -4.5    -12.9    5.0  2.06        3       13
#> log_lik[52]   -586.0   -517.0   -88.9   -434.5  197.1  2.06        3       13
#> log_lik[53]     -9.2     -8.5    -5.5     -8.0    1.4  2.06        3       13
#> log_lik[54]   -129.9    -80.0    -4.3    -74.6   46.6  2.06        3       13
#> log_lik[55]     -9.7     -7.9    -3.9     -7.5    2.2  2.06        3       13
#> log_lik[56]   -263.3   -238.7   -44.9   -199.5   87.7  2.06        3       13
#> log_lik[57]    -16.8    -15.4    -8.4    -14.1    3.3  2.06        3       13
#> log_lik[58]   -249.2   -151.2    -6.3   -141.5   90.0  2.06        3       13
#> log_lik[59]     -3.4     -3.3    -2.6     -3.2    0.3  1.87        4       13
#> log_lik[60]    -15.9    -10.1    -2.7    -10.0    4.8  2.06        3       13
#> log_lik[61]    -24.5    -22.2   -11.3    -20.3    5.2  2.06        3       13
#> log_lik[62]   -384.7   -232.9    -8.3   -217.8  139.5  2.06        3       13
#> log_lik[63]     -4.9     -4.6    -2.6     -4.2    0.9  2.06        4       13
#> log_lik[64]    -36.1    -23.4    -3.0    -22.0   12.3  2.06        3       13
#> log_lik[65]    -30.9    -28.0   -13.7    -25.4    6.7  2.06        3       13
#> log_lik[66]   -487.6   -294.8    -9.7   -275.7  177.2  2.06        3       13
#> log_lik[67]     -6.9     -6.4    -2.6     -5.7    1.7  2.06        4       13
#> log_lik[68]    -74.6    -50.3    -4.5    -45.9   26.3  2.06        3       13
#> log_lik[69]    -42.2    -38.1   -17.9    -34.4    9.5  2.06        3       13
#> log_lik[70]   -674.2   -405.1   -13.4   -379.9  244.9  2.06        3       13
#> log_lik[71]    -40.8    -34.7    -4.6    -29.3   14.4  2.06        3       13
#> log_lik[72]   -907.4   -718.6   -82.5   -619.2  316.0  2.06        3       13
#> log_lik[73]    -45.3    -40.9   -19.1    -36.9   10.3  2.06        3       13
#> log_lik[74]   -727.5   -436.6   -14.2   -409.6  264.3  2.06        3       13
#> log_lik[75]    -31.2    -26.8    -3.6    -22.6   10.9  2.06        4       13
#> log_lik[76]   -673.6   -522.5   -53.9   -452.4  236.4  2.06        3       13
#> log_lik[77]    -36.7    -33.2   -15.7    -30.0    8.2  2.06        3       13
#> log_lik[78]   -585.8   -350.6   -12.4   -329.6  212.5  2.06        3       13
#> log_lik[79]    -26.5    -22.6    -3.3    -19.2    9.2  2.06        4       13
#> log_lik[80]   -636.1   -496.4   -52.7   -429.1  222.8  2.06        3       13
#> psi[1]           0.5      0.5     0.5      0.5    0.0  1.05       10       13
#> psi[2]           0.4      0.4     0.4      0.4    0.0  1.06       12       13
#> xstar[1,1]      -7.6     -6.6    -4.3     -6.1    1.3  0.93       13       13
#> xstar[2,1]      -5.0     -3.5    -2.0     -3.4    1.1  0.95       13       13
#> sigma[1]         3.6      3.8     5.8      4.2    0.9  2.06        3       13
#> lp__        -14844.3 -11757.9 -1867.5 -10230.3 4939.7  2.06        3       13
#> 
#> For each parameter, Bulk_ESS and Tail_ESS are crude measures of 
#> effective sample size for bulk and tail quantities respectively (an ESS > 100 
#> per chain is considered good), and Rhat is the potential scale reduction 
#> factor on rank normalized split chains (at convergence, Rhat <= 1.05).

fitted <- dfa_fitted(m)
# }