Get the fitted values from a DFA as a data frame

Get the fitted values from a DFA as a data frame

dfa_fitted(modelfit, conf_level = 0.95, names = NULL)

Arguments

modelfit

Output from fit_dfa.

conf_level

Probability level for CI.

names

Optional vector of names for time series labels. Should be same length as the number of time series.

Value

A data frame with the following columns: ID is an identifier for each time series, time is the time step, y is the observed values standardized to mean 0 and unit variance, estimate is the mean fitted value, lower is the lower CI, and upper is the upper CI.

See also

Examples

# \donttest{
y <- sim_dfa(num_trends = 2, num_years = 20, num_ts = 4)
m <- fit_dfa(y = y$y_sim, num_trends = 2, iter = 50, chains = 1)
#> 
#> SAMPLING FOR MODEL 'dfa' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000121 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.21 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: WARNING: There aren't enough warmup iterations to fit the
#> Chain 1:          three stages of adaptation as currently configured.
#> Chain 1:          Reducing each adaptation stage to 15%/75%/10% of
#> Chain 1:          the given number of warmup iterations:
#> Chain 1:            init_buffer = 3
#> Chain 1:            adapt_window = 20
#> Chain 1:            term_buffer = 2
#> Chain 1: 
#> Chain 1: Iteration:  1 / 50 [  2%]  (Warmup)
#> Chain 1: Iteration:  5 / 50 [ 10%]  (Warmup)
#> Chain 1: Iteration: 10 / 50 [ 20%]  (Warmup)
#> Chain 1: Iteration: 15 / 50 [ 30%]  (Warmup)
#> Chain 1: Iteration: 20 / 50 [ 40%]  (Warmup)
#> Chain 1: Iteration: 25 / 50 [ 50%]  (Warmup)
#> Chain 1: Iteration: 26 / 50 [ 52%]  (Sampling)
#> Chain 1: Iteration: 30 / 50 [ 60%]  (Sampling)
#> Chain 1: Iteration: 35 / 50 [ 70%]  (Sampling)
#> Chain 1: Iteration: 40 / 50 [ 80%]  (Sampling)
#> Chain 1: Iteration: 45 / 50 [ 90%]  (Sampling)
#> Chain 1: Iteration: 50 / 50 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.008245 seconds (Warm-up)
#> Chain 1:                0.130133 seconds (Sampling)
#> Chain 1:                0.138378 seconds (Total)
#> Chain 1: 
#> Warning: There were 11 divergent transitions after warmup. See
#> https://mc-stan.org/misc/warnings.html#divergent-transitions-after-warmup
#> to find out why this is a problem and how to eliminate them.
#> Warning: There were 1 chains where the estimated Bayesian Fraction of Missing Information was low. See
#> https://mc-stan.org/misc/warnings.html#bfmi-low
#> Warning: Examine the pairs() plot to diagnose sampling problems
#> Warning: The largest R-hat is NA, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#r-hat
#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#bulk-ess
#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#tail-ess
#> Inference for the input samples (1 chains: each with iter = 25; warmup = 12):
#> 
#>                 Q5   Q50   Q95   Mean    SD  Rhat Bulk_ESS Tail_ESS
#> x[1,1]        -0.3   0.0   0.5    0.1   0.3  2.06        4       13
#> x[2,1]        -0.2   0.5   0.9    0.5   0.4  1.87        4       13
#> x[1,2]         0.9   1.0   1.3    1.1   0.2  1.58        4       13
#> x[2,2]         1.4   1.9   2.0    1.8   0.3  1.12        8       13
#> x[1,3]         0.6   0.8   0.9    0.8   0.1  1.45        5       13
#> x[2,3]         1.9   2.6   3.0    2.5   0.4  1.07       13       13
#> x[1,4]        -0.2   0.1   0.5    0.1   0.3  0.98       11       13
#> x[2,4]         2.6   3.4   3.7    3.3   0.4  1.71        4       13
#> x[1,5]         1.0   1.3   1.6    1.3   0.2  1.21        8       13
#> x[2,5]         2.6   3.4   3.9    3.3   0.5  2.06        4       13
#> x[1,6]        -0.2   0.1   0.3    0.1   0.2  0.95       13       13
#> x[2,6]         1.3   2.0   2.4    1.9   0.4  1.27        6       13
#> x[1,7]        -1.3  -0.7  -0.2   -0.8   0.4  1.45        5       13
#> x[2,7]         0.5   1.1   1.6    1.0   0.4  1.87        4       13
#> x[1,8]        -0.3  -0.1   0.3    0.0   0.2  1.19       13       13
#> x[2,8]        -0.2   0.1   0.7    0.1   0.3  1.25        5       13
#> x[1,9]         1.2   1.5   1.9    1.5   0.3  0.97       13       13
#> x[2,9]         0.1   0.6   1.2    0.6   0.4  1.39        5       13
#> x[1,10]       -0.1   0.5   1.0    0.5   0.5  1.15        7       13
#> x[2,10]        0.2   0.7   1.3    0.7   0.4  1.02       10       13
#> x[1,11]       -1.5  -1.1  -0.5   -1.0   0.4  1.11       10       13
#> x[2,11]       -0.2   0.4   0.8    0.4   0.4  1.06       10       13
#> x[1,12]       -0.7   0.1   0.7    0.2   0.6  1.24        6       13
#> x[2,12]       -1.5  -0.6  -0.2   -0.8   0.5  1.58        4       13
#> x[1,13]       -1.7  -0.7  -0.1   -0.8   0.6  1.19        6       13
#> x[2,13]        0.4   1.1   1.6    1.0   0.5  1.58        9       13
#> x[1,14]       -2.6  -1.3  -0.2   -1.4   0.8  1.87        4       13
#> x[2,14]        0.0   0.8   1.2    0.6   0.5  1.09        9       13
#> x[1,15]       -2.6  -1.9  -0.7   -1.8   0.7  1.37        5       13
#> x[2,15]       -0.9   0.0   0.8   -0.1   0.6  1.11        7       13
#> x[1,16]       -1.3  -0.3   1.3   -0.1   1.0  1.58        4       13
#> x[2,16]        0.1   1.6   1.9    1.2   0.8  1.71        4       13
#> x[1,17]       -1.5  -0.8   0.5   -0.5   0.8  1.32        5       13
#> x[2,17]       -1.1   0.6   1.0    0.1   0.8  1.58        4       13
#> x[1,18]       -1.2  -0.7   0.6   -0.3   0.8  1.32        5       13
#> x[2,18]       -1.7  -0.2   0.2   -0.6   0.7  1.38        6       13
#> x[1,19]       -2.7  -1.8  -0.3   -1.5   0.9  1.47        4       13
#> x[2,19]       -2.2  -1.2  -0.5   -1.3   0.6  1.32       11       13
#> x[1,20]       -2.0  -1.0   0.6   -0.7   1.0  1.39        4       13
#> x[2,20]       -3.1  -2.5  -1.7   -2.4   0.5  1.15       13       13
#> Z[1,1]        -0.4   0.0   0.9    0.1   0.5  1.58       13       13
#> Z[2,1]        -1.0   0.3   0.7    0.2   0.6  1.13        7       13
#> Z[3,1]        -0.8  -0.5   1.2   -0.1   0.8  1.71       13       13
#> Z[4,1]        -1.3   0.3   0.7    0.0   0.8  1.71       11       13
#> Z[1,2]         0.0   0.0   0.0    0.0   0.0  1.00       13       13
#> Z[2,2]        -1.2   0.1   3.5    0.4   1.6  1.71       10       13
#> Z[3,2]        -1.5  -0.1   1.2   -0.1   0.9  1.58       10       13
#> Z[4,2]        -0.1   0.3   0.5    0.2   0.2  1.16       11       13
#> log_lik[1]    -5.8  -4.6  -2.7   -4.6   1.0  1.15       13       13
#> log_lik[2]    -4.9  -0.8  -0.6   -1.9   1.7  1.87        5       13
#> log_lik[3]    -4.9  -2.6  -2.1   -3.0   1.0  1.30        5       13
#> log_lik[4]    -5.0  -3.6  -2.5   -3.6   0.9  1.06       11       13
#> log_lik[5]    -4.9  -3.3  -2.3   -3.4   0.9  1.06        7       13
#> log_lik[6]    -4.9  -1.2  -0.8   -2.1   1.6  2.06        6       13
#> log_lik[7]    -4.9  -1.3  -0.6   -2.0   1.6  1.87        4       13
#> log_lik[8]    -4.9  -1.4  -0.9   -2.2   1.5  1.87        4       13
#> log_lik[9]    -4.9  -1.0  -0.8   -2.0   1.6  1.87        4       13
#> log_lik[10]   -4.9  -0.8  -0.6   -1.9   1.7  1.58        5       13
#> log_lik[11]   -4.9  -1.3  -0.7   -2.1   1.6  2.06        6       13
#> log_lik[12]   -4.9  -1.2  -0.8   -2.2   1.5  1.87        6       13
#> log_lik[13]   -4.9  -0.7  -0.6   -1.8   1.7  1.87        5       13
#> log_lik[14]   -4.9  -1.8  -0.8   -2.3   1.6  2.06        4       13
#> log_lik[15]   -4.9  -1.2  -0.6   -2.1   1.7  1.87        4       13
#> log_lik[16]   -4.9  -1.0  -0.6   -1.9   1.7  1.71        5       13
#> log_lik[17]   -4.9  -0.8  -0.6   -1.9   1.7  1.87        4       13
#> log_lik[18]   -4.9  -0.8  -0.6   -1.9   1.8  1.87        5       13
#> log_lik[19]   -4.9  -1.8  -0.7   -2.3   1.6  2.06        6       13
#> log_lik[20]   -4.9  -1.7  -1.1   -2.4   1.4  1.38        7       13
#> log_lik[21]   -4.9  -1.5  -1.3   -2.3   1.4  1.87        4       13
#> log_lik[22]   -4.9  -2.3  -1.2   -2.6   1.3  2.06        4       13
#> log_lik[23]   -4.9  -0.9  -0.6   -1.9   1.7  1.58        4       13
#> log_lik[24]   -4.9  -0.9  -0.6   -1.9   1.7  2.06        6       13
#> log_lik[25]   -4.9  -0.8  -0.6   -1.9   1.7  1.87        5       13
#> log_lik[26]   -4.9  -2.8  -1.7   -2.9   1.1  1.87        4       13
#> log_lik[27]   -4.9  -1.8  -1.4   -2.5   1.3  2.06        4       13
#> log_lik[28]   -4.9  -1.3  -0.9   -2.2   1.5  2.06        4       13
#> log_lik[29]   -4.9  -0.7  -0.6   -1.8   1.7  1.87        5       13
#> log_lik[30]   -4.9  -1.2  -0.8   -2.1   1.6  1.87        5       13
#> log_lik[31]   -4.9  -0.9  -0.6   -1.9   1.7  2.06        6       13
#> log_lik[32]   -4.9  -0.8  -0.6   -1.9   1.7  2.06        6       13
#> log_lik[33]   -4.9  -0.9  -0.6   -1.9   1.7  1.71        5       13
#> log_lik[34]   -4.9  -0.9  -0.6   -1.9   1.7  1.87        4       13
#> log_lik[35]   -4.9  -1.4  -0.6   -2.0   1.7  1.58        4       13
#> log_lik[36]   -4.9  -0.9  -0.6   -1.9   1.7  1.58        4       13
#> log_lik[37]   -4.9  -1.9  -1.2   -2.4   1.3  2.06        4       13
#> log_lik[38]   -4.9  -1.1  -0.7   -2.0   1.6  1.71        5       13
#> log_lik[39]   -4.9  -0.9  -0.6   -1.9   1.7  2.06        6       13
#> log_lik[40]   -4.9  -1.2  -0.7   -2.0   1.6  1.87        4       13
#> log_lik[41]   -4.9  -1.0  -0.7   -1.9   1.7  1.87        5       13
#> log_lik[42]   -4.9  -0.9  -0.7   -1.9   1.7  1.87        4       13
#> log_lik[43]   -4.9  -1.1  -0.7   -2.0   1.6  1.71        5       13
#> log_lik[44]   -4.9  -0.7  -0.6   -1.9   1.7  1.87        5       13
#> log_lik[45]   -4.9  -1.2  -0.9   -2.1   1.5  2.06        4       13
#> log_lik[46]   -4.9  -1.0  -0.7   -1.9   1.7  1.87        5       13
#> log_lik[47]   -4.9  -1.3  -0.6   -2.0   1.6  2.06        4       13
#> log_lik[48]   -4.9  -1.3  -0.6   -2.0   1.6  2.06        4       13
#> log_lik[49]   -4.9  -0.9  -0.6   -1.9   1.7  2.06        6       13
#> log_lik[50]   -4.9  -0.8  -0.6   -1.9   1.7  2.06        6       13
#> log_lik[51]   -4.9  -1.0  -0.8   -2.1   1.6  1.71        5       13
#> log_lik[52]   -4.9  -1.0  -0.7   -2.0   1.7  1.87        5       13
#> log_lik[53]   -4.9  -2.8  -1.3   -2.9   1.2  1.25        5       13
#> log_lik[54]   -4.9  -1.4  -1.1   -2.3   1.4  1.87        6       13
#> log_lik[55]   -4.9  -0.9  -0.6   -2.0   1.7  1.71        5       13
#> log_lik[56]   -4.9  -1.1  -0.7   -2.0   1.7  1.71        5       13
#> log_lik[57]   -4.9  -2.1  -0.8   -2.4   1.4  1.71        4       13
#> log_lik[58]   -4.9  -1.0  -0.7   -2.0   1.6  2.06        6       13
#> log_lik[59]   -4.9  -0.7  -0.6   -1.9   1.8  1.87        5       13
#> log_lik[60]   -4.9  -0.8  -0.6   -1.9   1.7  1.87        5       13
#> log_lik[61]   -4.9  -2.2  -1.7   -2.7   1.2  1.39        4       13
#> log_lik[62]   -4.9  -2.5  -1.4   -2.8   1.2  1.47        4       13
#> log_lik[63]   -4.9  -1.4  -0.7   -2.1   1.6  1.87        4       13
#> log_lik[64]   -4.9  -1.0  -0.6   -1.9   1.7  1.87        4       13
#> log_lik[65]   -4.9  -0.8  -0.6   -1.9   1.7  1.87        5       13
#> log_lik[66]   -4.9  -1.7  -0.7   -2.1   1.6  1.87        4       13
#> log_lik[67]   -4.9  -1.6  -0.6   -2.1   1.6  1.87        4       13
#> log_lik[68]   -4.9  -1.3  -0.6   -2.0   1.6  1.87        4       13
#> log_lik[69]   -4.9  -1.0  -0.7   -2.0   1.6  1.87        4       13
#> log_lik[70]   -4.9  -1.9  -0.6   -2.1   1.6  2.06        4       13
#> log_lik[71]   -4.9  -1.7  -0.7   -2.2   1.6  2.06        4       13
#> log_lik[72]   -4.9  -1.6  -0.6   -2.1   1.6  2.06        4       13
#> log_lik[73]   -4.9  -2.5  -1.3   -2.8   1.2  1.10        5       13
#> log_lik[74]   -4.9  -0.9  -0.7   -2.0   1.7  1.87        6       13
#> log_lik[75]   -4.9  -1.5  -0.7   -2.1   1.6  2.06        4       13
#> log_lik[76]   -4.9  -2.4  -0.8   -2.3   1.5  2.06        4       13
#> log_lik[77]   -4.9  -2.1  -1.2   -2.6   1.3  1.71        4       13
#> log_lik[78]   -4.9  -1.6  -0.7   -2.1   1.6  2.06        4       13
#> log_lik[79]   -4.9  -2.1  -0.8   -2.5   1.5  2.06        4       13
#> log_lik[80]   -4.9  -2.3  -0.8   -2.6   1.6  2.06        4       13
#> xstar[1,1]    -2.8  -0.6   1.7   -0.4   1.7  1.25        6       13
#> xstar[2,1]    -3.5  -2.7  -1.5   -2.6   0.7  1.30        6       13
#> sigma[1]       0.7   0.8  55.2   11.9  23.3  1.87        5       13
#> lp__        -401.3 -98.5 -77.7 -171.1 128.6  2.06        3       13
#> 
#> For each parameter, Bulk_ESS and Tail_ESS are crude measures of 
#> effective sample size for bulk and tail quantities respectively (an ESS > 100 
#> per chain is considered good), and Rhat is the potential scale reduction 
#> factor on rank normalized split chains (at convergence, Rhat <= 1.05).
fitted <- dfa_fitted(m)
# }