Get the fitted values from a DFA as a data frame

Get the fitted values from a DFA as a data frame

dfa_fitted(modelfit, conf_level = 0.95, names = NULL)

Arguments

modelfit

Output from fit_dfa.

conf_level

Probability level for CI.

names

Optional vector of names for time series labels. Should be same length as the number of time series.

Value

A data frame with the following columns: ID is an identifier for each time series, time is the time step, y is the observed values standardized to mean 0 and unit variance, estimate is the mean fitted value, lower is the lower CI, and upper is the upper CI.

See also

Examples

# \donttest{
y <- sim_dfa(num_trends = 2, num_years = 20, num_ts = 4)
m <- fit_dfa(y = y$y_sim, num_trends = 2, iter = 50, chains = 1)
#> 
#> SAMPLING FOR MODEL 'dfa' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000202 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 2.02 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: WARNING: There aren't enough warmup iterations to fit the
#> Chain 1:          three stages of adaptation as currently configured.
#> Chain 1:          Reducing each adaptation stage to 15%/75%/10% of
#> Chain 1:          the given number of warmup iterations:
#> Chain 1:            init_buffer = 3
#> Chain 1:            adapt_window = 20
#> Chain 1:            term_buffer = 2
#> Chain 1: 
#> Chain 1: Iteration:  1 / 50 [  2%]  (Warmup)
#> Chain 1: Iteration:  5 / 50 [ 10%]  (Warmup)
#> Chain 1: Iteration: 10 / 50 [ 20%]  (Warmup)
#> Chain 1: Iteration: 15 / 50 [ 30%]  (Warmup)
#> Chain 1: Iteration: 20 / 50 [ 40%]  (Warmup)
#> Chain 1: Iteration: 25 / 50 [ 50%]  (Warmup)
#> Chain 1: Iteration: 26 / 50 [ 52%]  (Sampling)
#> Chain 1: Iteration: 30 / 50 [ 60%]  (Sampling)
#> Chain 1: Iteration: 35 / 50 [ 70%]  (Sampling)
#> Chain 1: Iteration: 40 / 50 [ 80%]  (Sampling)
#> Chain 1: Iteration: 45 / 50 [ 90%]  (Sampling)
#> Chain 1: Iteration: 50 / 50 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.015988 seconds (Warm-up)
#> Chain 1:                0.019211 seconds (Sampling)
#> Chain 1:                0.035199 seconds (Total)
#> Chain 1: 
#> Warning: There were 1 chains where the estimated Bayesian Fraction of Missing Information was low. See
#> https://mc-stan.org/misc/warnings.html#bfmi-low
#> Warning: Examine the pairs() plot to diagnose sampling problems
#> Warning: The largest R-hat is NA, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#r-hat
#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#bulk-ess
#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#tail-ess
#> Inference for the input samples (1 chains: each with iter = 25; warmup = 12):
#> 
#>                   Q5      Q50      Q95     Mean     SD  Rhat Bulk_ESS Tail_ESS
#> x[1,1]          -1.3     -1.3     -1.3     -1.3    0.0  1.87        4       13
#> x[2,1]           0.0      0.0      0.0      0.0    0.0  2.06        3       13
#> x[1,2]          -0.1     -0.1     -0.1     -0.1    0.0  2.06        4       13
#> x[2,2]           1.2      1.2      1.2      1.2    0.0  2.06        4       13
#> x[1,3]          -1.5     -1.5     -1.5     -1.5    0.0  2.06        3       13
#> x[2,3]          -0.5     -0.4     -0.4     -0.4    0.0  2.06        3       13
#> x[1,4]          -1.3     -1.3     -1.3     -1.3    0.0  2.06        3       13
#> x[2,4]          -1.7     -1.7     -1.7     -1.7    0.0  2.06        3       13
#> x[1,5]          -2.3     -2.3     -2.3     -2.3    0.0  2.06        3       13
#> x[2,5]          -1.2     -1.2     -1.2     -1.2    0.0  2.06        3       13
#> x[1,6]          -3.9     -3.9     -3.9     -3.9    0.0  2.06        3       13
#> x[2,6]          -0.1     -0.1     -0.1     -0.1    0.0  2.06        3       13
#> x[1,7]          -2.3     -2.2     -2.2     -2.2    0.0  2.06        3       13
#> x[2,7]          -1.4     -1.4     -1.3     -1.4    0.0  2.06        3       13
#> x[1,8]          -2.4     -2.4     -2.3     -2.4    0.0  2.06        3       13
#> x[2,8]          -1.0     -1.0     -0.9     -1.0    0.0  2.06        3       13
#> x[1,9]          -2.0     -2.0     -1.9     -2.0    0.0  2.06        3       13
#> x[2,9]          -0.4     -0.4     -0.3     -0.3    0.0  2.06        3       13
#> x[1,10]         -1.7     -1.6     -1.6     -1.6    0.0  2.06        3       13
#> x[2,10]          0.7      0.7      0.8      0.7    0.0  2.06        3       13
#> x[1,11]         -0.3     -0.2     -0.2     -0.2    0.0  2.06        3       13
#> x[2,11]          0.6      0.6      0.7      0.6    0.0  2.06        3       13
#> x[1,12]          1.3      1.3      1.4      1.3    0.0  2.06        3       13
#> x[2,12]         -0.8     -0.8     -0.7     -0.8    0.0  2.06        3       13
#> x[1,13]          0.7      0.8      0.8      0.7    0.1  2.06        3       13
#> x[2,13]          0.2      0.3      0.4      0.3    0.1  2.06        3       13
#> x[1,14]          1.8      1.9      1.9      1.9    0.1  2.06        3       13
#> x[2,14]         -0.4     -0.4     -0.2     -0.3    0.1  2.06        3       13
#> x[1,15]          2.5      2.6      2.7      2.6    0.1  2.06        3       13
#> x[2,15]         -2.0     -1.9     -1.8     -1.9    0.1  2.06        3       13
#> x[1,16]          4.0      4.1      4.2      4.1    0.1  2.06        3       13
#> x[2,16]         -1.7     -1.6     -1.5     -1.6    0.1  2.06        3       13
#> x[1,17]          5.3      5.5      5.5      5.4    0.1  2.06        3       13
#> x[2,17]         -3.5     -3.4     -3.3     -3.4    0.1  2.06        3       13
#> x[1,18]          6.3      6.4      6.5      6.4    0.1  2.06        3       13
#> x[2,18]         -4.9     -4.8     -4.7     -4.8    0.1  2.06        3       13
#> x[1,19]          5.8      5.9      6.0      5.9    0.1  2.06        3       13
#> x[2,19]         -4.1     -4.1     -3.9     -4.1    0.1  2.06        3       13
#> x[1,20]          3.9      4.0      4.1      4.0    0.1  2.06        3       13
#> x[2,20]         -5.3     -5.2     -5.1     -5.2    0.1  2.06        3       13
#> Z[1,1]         -98.0    -98.0    -98.0    -98.0    0.0  2.06        4       13
#> Z[2,1]          24.0     24.7     25.1     24.7    0.4  2.06        3       13
#> Z[3,1]         -28.3    -28.1    -27.9    -28.1    0.2  2.06        3       13
#> Z[4,1]         -48.4    -47.9    -47.2    -47.9    0.4  2.06        3       13
#> Z[1,2]           0.0      0.0      0.0      0.0    0.0  1.00       13       13
#> Z[2,2]         -96.7    -96.7    -96.7    -96.7    0.0  1.58        4       13
#> Z[3,2]         -51.4    -51.3    -50.9    -51.2    0.2  2.06        4       13
#> Z[4,2]          29.8     30.2     30.4     30.2    0.2  2.06        3       13
#> log_lik[1]     -57.2    -52.1    -45.5    -52.2    4.3  2.06        3       13
#> log_lik[2]      -7.8     -7.5     -6.9     -7.4    0.3  2.06        3       13
#> log_lik[3]      -7.6     -7.1     -6.6     -7.1    0.4  2.06        3       13
#> log_lik[4]     -17.2    -15.7    -13.9    -15.8    1.2  2.06        3       13
#> log_lik[5]      -4.0     -3.9     -3.9     -3.9    0.0  2.06        4       13
#> log_lik[6]     -54.5    -50.1    -44.5    -50.2    3.7  2.06        3       13
#> log_lik[7]     -15.4    -14.2    -12.8    -14.3    1.0  2.06        3       13
#> log_lik[8]     -10.0     -9.5     -8.8     -9.5    0.5  2.06        3       13
#> log_lik[9]     -76.9    -70.4    -61.9    -70.5    5.5  2.06        3       13
#> log_lik[10]     -3.6     -3.6     -3.5     -3.6    0.0  1.71        4       13
#> log_lik[11]    -18.5    -16.7    -14.4    -16.8    1.5  2.06        3       13
#> log_lik[12]    -15.0    -13.9    -12.7    -14.0    0.8  2.06        3       13
#> log_lik[13]    -58.5    -54.1    -48.5    -54.2    3.7  2.06        3       13
#> log_lik[14]    -68.1    -60.2    -50.5    -60.6    6.5  2.06        3       13
#> log_lik[15]    -59.2    -52.8    -44.6    -53.0    5.4  2.06        3       13
#> log_lik[16]     -3.8     -3.7     -3.6     -3.7    0.1  2.06        3       13
#> log_lik[17]   -173.3   -158.9   -140.3   -159.2   12.1  2.06        3       13
#> log_lik[18]    -16.4    -14.5    -12.2    -14.6    1.5  2.06        3       13
#> log_lik[19]    -59.9    -53.4    -44.9    -53.5    5.5  2.06        3       13
#> log_lik[20]    -21.1    -19.6    -18.1    -19.7    1.1  2.06        3       13
#> log_lik[21]   -496.1   -451.9   -394.8   -452.9   37.3  2.06        3       13
#> log_lik[22]    -28.9    -27.0    -24.7    -27.1    1.6  2.06        3       13
#> log_lik[23]    -49.5    -44.0    -37.0    -44.2    4.6  2.06        3       13
#> log_lik[24]   -119.9   -108.2    -93.8   -108.6    9.6  2.06        3       13
#> log_lik[25]   -161.1   -148.6   -132.6   -149.0   10.5  2.06        3       13
#> log_lik[26]    -25.7    -22.5    -18.1    -22.5    2.8  2.06        3       13
#> log_lik[27]    -65.6    -58.4    -48.9    -58.6    6.2  2.06        3       13
#> log_lik[28]    -17.6    -16.7    -15.8    -16.8    0.7  2.06        3       13
#> log_lik[29]   -184.5   -170.5   -152.2   -170.9   11.9  2.06        3       13
#> log_lik[30]     -8.1     -7.1     -5.9     -7.2    0.8  2.06        3       13
#> log_lik[31]    -51.0    -45.3    -37.9    -45.5    4.8  2.06        3       13
#> log_lik[32]    -26.8    -25.2    -23.4    -25.3    1.3  2.06        3       13
#> log_lik[33]   -129.5   -120.7   -109.2   -121.0    7.5  2.06        3       13
#> log_lik[34]     -4.5     -4.0     -3.8     -4.1    0.3  2.06        3       13
#> log_lik[35]    -23.3    -20.8    -17.5    -20.9    2.1  2.06        3       13
#> log_lik[36]    -26.0    -24.7    -23.1    -24.8    1.1  2.06        3       13
#> log_lik[37]    -88.2    -83.2    -77.0    -83.6    4.1  2.06        3       13
#> log_lik[38]    -43.8    -42.5    -41.3    -42.6    0.9  2.06        3       13
#> log_lik[39]     -3.7     -3.7     -3.6     -3.7    0.0  2.06        4       13
#> log_lik[40]    -37.0    -34.9    -32.5    -35.1    1.7  2.06        3       13
#> log_lik[41]     -5.5     -4.7     -4.3     -4.8    0.4  2.06        3       13
#> log_lik[42]    -17.6    -15.9    -15.3    -16.1    0.8  2.06        3       13
#> log_lik[43]     -5.4     -5.4     -5.4     -5.4    0.0  0.98        9       13
#> log_lik[44]     -6.6     -6.0     -5.6     -6.0    0.3  2.06        3       13
#> log_lik[45]    -68.0    -57.7    -44.9    -57.8    8.6  2.06        3       13
#> log_lik[46]    -49.1    -40.4    -28.9    -40.5    7.4  2.06        3       13
#> log_lik[47]     -3.5     -3.5     -3.4     -3.5    0.0  2.06        3       13
#> log_lik[48]    -33.5    -28.1    -21.2    -28.2    4.5  2.06        3       13
#> log_lik[49]    -25.1    -20.7    -15.1    -20.7    3.7  2.06        3       13
#> log_lik[50]     -4.6     -3.6     -3.4     -3.8    0.4  2.06        3       13
#> log_lik[51]     -7.7     -7.5     -7.4     -7.6    0.1  2.06        4       13
#> log_lik[52]     -6.9     -5.8     -4.6     -5.9    0.8  2.06        3       13
#> log_lik[53]   -126.4   -107.7    -83.8   -107.8   15.7  2.06        3       13
#> log_lik[54]    -29.5    -23.4    -15.2    -23.4    5.3  2.06        3       13
#> log_lik[55]     -7.5     -7.3     -7.2     -7.3    0.1  2.06        4       13
#> log_lik[56]    -41.4    -34.4    -25.4    -34.4    5.9  2.06        3       13
#> log_lik[57]   -247.4   -212.9   -168.9   -213.3   28.9  2.06        3       13
#> log_lik[58]   -235.2   -200.0   -152.2   -200.1   30.5  2.06        3       13
#> log_lik[59]     -5.5     -5.2     -4.7     -5.2    0.3  2.06        3       13
#> log_lik[60]   -129.0   -109.4    -83.8   -109.5   16.6  2.06        3       13
#> log_lik[61]   -588.4   -512.8   -416.1   -513.9   63.4  2.06        3       13
#> log_lik[62]   -250.0   -211.4   -159.0   -211.7   33.4  2.06        3       13
#> log_lik[63]     -7.1     -6.9     -6.8     -7.0    0.1  1.87        4       13
#> log_lik[64]   -225.0   -191.9   -149.2   -192.3   27.8  2.06        3       13
#> log_lik[65]  -1014.2   -889.4   -729.5   -891.7  104.7  2.06        3       13
#> log_lik[66]   -780.8   -675.1   -533.0   -676.3   91.0  2.06        3       13
#> log_lik[67]     -5.0     -4.8     -4.4     -4.8    0.2  2.06        4       13
#> log_lik[68]   -484.2   -417.4   -330.4   -418.3   56.4  2.06        3       13
#> log_lik[69]  -1403.0  -1234.1  -1017.9  -1237.4  141.7  2.06        3       13
#> log_lik[70]  -1408.6  -1228.1   -985.3  -1230.0  155.5  2.06        3       13
#> log_lik[71]    -19.3    -17.5    -14.2    -17.4    1.8  2.06        3       13
#> log_lik[72]   -747.2   -647.2   -516.2   -648.3   84.7  2.06        3       13
#> log_lik[73]  -1193.6  -1047.2   -859.8  -1050.0  122.8  2.06        3       13
#> log_lik[74]  -1040.8   -902.9   -717.3   -904.5  118.8  2.06        3       13
#> log_lik[75]     -9.2     -8.6     -7.2     -8.5    0.7  2.06        3       13
#> log_lik[76]   -594.8   -513.3   -407.0   -514.4   68.8  2.06        3       13
#> log_lik[77]   -553.7   -479.1   -385.3   -480.8   62.0  2.06        3       13
#> log_lik[78]  -1285.3  -1121.1   -900.7  -1123.1  141.1  2.06        3       13
#> log_lik[79]    -84.8    -75.9    -61.8    -75.6    8.4  2.06        3       13
#> log_lik[80]   -440.0   -378.4   -297.1   -378.9   52.3  2.06        3       13
#> xstar[1,1]       2.3      3.9      5.4      3.8    1.1  0.98       13       13
#> xstar[2,1]      -7.1     -5.3     -3.8     -5.2    1.2  0.91       13       13
#> sigma[1]        12.1     12.7     13.7     12.7    0.6  2.06        3       13
#> lp__        -28676.7 -26708.9 -24112.0 -26733.3 1676.3  2.06        3       13
#> 
#> For each parameter, Bulk_ESS and Tail_ESS are crude measures of 
#> effective sample size for bulk and tail quantities respectively (an ESS > 100 
#> per chain is considered good), and Rhat is the potential scale reduction 
#> factor on rank normalized split chains (at convergence, Rhat <= 1.05).
fitted <- dfa_fitted(m)
# }