Find the best number of trends according to LOOIC

Fit a DFA with different number of trends and return the leave one out (LOO) value as calculated by the loo package.

find_dfa_trends(
  y = y,
  kmin = 1,
  kmax = 5,
  iter = 2000,
  thin = 1,
  compare_normal = FALSE,
  convergence_threshold = 1.05,
  variance = c("equal", "unequal"),
  ...
)

Arguments

y
A matrix of data to fit. Columns represent time element.

kmin
Minimum number of trends, defaults to 1.

kmax
Maximum number of trends, defaults to 5.

iter
Iterations when sampling from each Stan model, defaults to 2000.

thin
Thinning rate when sampling from each Stan model, defaults to 1.

compare_normal
If TRUE, does model selection comparison of Normal vs. Student-t errors

convergence_threshold
The maximum allowed value of Rhat to determine convergence of parameters

variance
Vector of variance arguments for searching over large groups of models. Can be either or both of ("equal","unequal")

...
Other arguments to pass to fit_dfa()

Examples

# \donttest{
set.seed(42)
s <- sim_dfa(num_trends = 2, num_years = 20, num_ts = 3)
# only 1 chain and 180 iterations used so example runs quickly:
m <- find_dfa_trends(
  y = s$y_sim, iter = 50,
  kmin = 1, kmax = 2, chains = 1, compare_normal = FALSE,
  variance = "equal", convergence_threshold = 1.1,
  control = list(adapt_delta = 0.95, max_treedepth = 20))

#> 
#> SAMPLING FOR MODEL 'dfa' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 4.2e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.42 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: WARNING: There aren't enough warmup iterations to fit the
#> Chain 1:          three stages of adaptation as currently configured.
#> Chain 1:          Reducing each adaptation stage to 15%/75%/10% of
#> Chain 1:          the given number of warmup iterations:
#> Chain 1:            init_buffer = 3
#> Chain 1:            adapt_window = 20
#> Chain 1:            term_buffer = 2
#> Chain 1: 
#> Chain 1: Iteration:  1 / 50 [  2%]  (Warmup)
#> Chain 1: Iteration:  5 / 50 [ 10%]  (Warmup)
#> Chain 1: Iteration: 10 / 50 [ 20%]  (Warmup)
#> Chain 1: Iteration: 15 / 50 [ 30%]  (Warmup)
#> Chain 1: Iteration: 20 / 50 [ 40%]  (Warmup)
#> Chain 1: Iteration: 25 / 50 [ 50%]  (Warmup)
#> Chain 1: Iteration: 26 / 50 [ 52%]  (Sampling)
#> Chain 1: Iteration: 30 / 50 [ 60%]  (Sampling)
#> Chain 1: Iteration: 35 / 50 [ 70%]  (Sampling)
#> Chain 1: Iteration: 40 / 50 [ 80%]  (Sampling)
#> Chain 1: Iteration: 45 / 50 [ 90%]  (Sampling)
#> Chain 1: Iteration: 50 / 50 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 1.53544 seconds (Warm-up)
#> Chain 1:                1.6628 seconds (Sampling)
#> Chain 1:                3.19824 seconds (Total)
#> Chain 1: 

#> Warning: The largest R-hat is 1.4, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#r-hat

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#tail-ess

#> Inference for the input samples (1 chains: each with iter = 25; warmup = 12):
#> 
#>                Q5   Q50   Q95  Mean   SD  Rhat Bulk_ESS Tail_ESS
#> x[1,1]        0.7   2.3   4.1   2.4  1.2  0.94       13       13
#> x[1,2]        0.7   2.0   3.8   2.1  1.1  0.95       13       13
#> x[1,3]        0.5   1.0   2.8   1.3  0.9  1.05       13       13
#> x[1,4]       -1.3  -0.4   0.8  -0.4  0.8  0.99       13       13
#> x[1,5]       -3.9  -2.7  -0.6  -2.3  1.3  0.92       13       13
#> x[1,6]       -4.6  -3.0  -1.1  -2.9  1.3  0.97       13       13
#> x[1,7]       -4.3  -3.0  -1.6  -2.9  0.9  0.92       13       13
#> x[1,8]       -5.8  -3.0  -1.2  -3.3  1.7  1.12       13       13
#> x[1,9]       -6.3  -3.7  -1.4  -3.7  1.7  0.92       13       13
#> x[1,10]      -6.0  -3.7  -1.2  -3.6  1.7  1.01       13       13
#> x[1,11]      -4.5  -2.2  -0.5  -2.3  1.4  1.07       13       13
#> x[1,12]      -3.2  -1.1  -0.1  -1.4  1.1  1.04       13       13
#> x[1,13]      -0.3   0.8   1.6   0.7  0.7  0.93       13       13
#> x[1,14]       1.4   2.4   6.2   3.1  1.8  0.98       13       13
#> x[1,15]       2.0   3.3   6.1   3.7  1.6  0.92       13       13
#> x[1,16]       1.5   3.1   6.6   3.6  1.9  0.95       13       13
#> x[1,17]       0.9   2.7   5.3   2.8  1.6  0.99       13       13
#> x[1,18]       0.3   1.0   3.7   1.4  1.3  0.92       13       13
#> x[1,19]      -0.7   0.9   3.1   1.1  1.4  0.98       13       13
#> x[1,20]       0.5   2.5   4.9   2.5  1.7  0.92       13       13
#> Z[1,1]        0.2   0.2   0.5   0.3  0.1  0.98       13       13
#> Z[2,1]        0.2   0.3   0.8   0.4  0.2  0.93       13       13
#> Z[3,1]        0.2   0.4   0.7   0.4  0.2  0.92       13       13
#> log_lik[1]   -3.3  -2.1  -1.4  -2.1  0.7  1.07       13       13
#> log_lik[2]   -1.7  -1.1  -0.7  -1.2  0.4  1.32       13       13
#> log_lik[3]   -1.8  -0.9  -0.6  -1.1  0.4  1.07        9       13
#> log_lik[4]   -1.8  -0.6  -0.4  -0.8  0.5  1.01       13       13
#> log_lik[5]   -1.6  -1.2  -0.6  -1.1  0.4  0.93       13       13
#> log_lik[6]   -1.6  -1.1  -0.3  -0.9  0.5  1.15       13       13
#> log_lik[7]   -4.6  -2.4  -2.0  -2.8  0.9  1.01       13       13
#> log_lik[8]   -1.0  -0.7  -0.4  -0.7  0.2  1.09        8       13
#> log_lik[9]   -0.6  -0.5  -0.3  -0.5  0.1  0.94       13       13
#> log_lik[10]  -1.2  -0.9  -0.4  -0.9  0.3  0.92       13       13
#> log_lik[11]  -0.7  -0.4  -0.3  -0.4  0.1  1.24        8       13
#> log_lik[12]  -0.7  -0.5  -0.3  -0.5  0.1  1.10       13       13
#> log_lik[13]  -2.1  -1.0  -0.7  -1.2  0.5  0.99       13       13
#> log_lik[14]  -1.0  -0.4  -0.2  -0.5  0.3  1.11       13       13
#> log_lik[15]  -1.5  -0.5  -0.3  -0.7  0.5  1.08       13       13
#> log_lik[16]  -1.4  -0.6  -0.4  -0.8  0.4  0.97       13       13
#> log_lik[17]  -1.6  -0.8  -0.3  -0.9  0.5  1.09       13       13
#> log_lik[18]  -0.7  -0.4  -0.2  -0.4  0.2  1.09       12       13
#> log_lik[19]  -3.2  -1.8  -1.4  -2.1  0.7  1.18        9       13
#> log_lik[20]  -1.4  -0.7  -0.4  -0.7  0.4  1.04       13       13
#> log_lik[21]  -1.1  -0.5  -0.3  -0.6  0.2  1.32        7       13
#> log_lik[22]  -0.9  -0.4  -0.3  -0.5  0.2  0.95       13       13
#> log_lik[23]  -1.6  -0.4  -0.3  -0.6  0.5  0.96       12       13
#> log_lik[24]  -1.0  -0.5  -0.3  -0.5  0.3  0.91       13       13
#> log_lik[25]  -1.3  -0.6  -0.4  -0.8  0.4  1.10       13       13
#> log_lik[26]  -0.7  -0.5  -0.3  -0.5  0.2  1.71        4       13
#> log_lik[27]  -0.8  -0.3  -0.2  -0.4  0.2  1.30        7       13
#> log_lik[28]  -1.2  -0.4  -0.3  -0.6  0.3  1.37        5       13
#> log_lik[29]  -0.8  -0.4  -0.3  -0.4  0.2  1.13       10       13
#> log_lik[30]  -0.8  -0.4  -0.3  -0.5  0.2  1.27        6       13
#> log_lik[31]  -1.9  -1.1  -0.5  -1.1  0.6  0.96       13       13
#> log_lik[32]  -1.7  -1.0  -0.5  -1.1  0.4  1.00       13       13
#> log_lik[33]  -1.5  -0.8  -0.5  -0.8  0.4  0.91       13       13
#> log_lik[34]  -1.9  -1.4  -0.6  -1.3  0.5  0.98       13       13
#> log_lik[35]  -0.9  -0.5  -0.3  -0.5  0.2  1.32       13       13
#> log_lik[36]  -0.7  -0.4  -0.3  -0.4  0.1  1.30        6       13
#> log_lik[37]  -0.9  -0.5  -0.3  -0.6  0.2  0.93       13       13
#> log_lik[38]  -1.1  -0.6  -0.4  -0.7  0.3  1.24        8       13
#> log_lik[39]  -0.9  -0.5  -0.3  -0.6  0.2  1.47        5       13
#> log_lik[40]  -0.8  -0.4  -0.3  -0.5  0.2  1.45        4       13
#> log_lik[41]  -1.1  -0.4  -0.3  -0.5  0.3  1.87        4       13
#> log_lik[42]  -1.0  -0.6  -0.4  -0.6  0.2  1.06       13       13
#> log_lik[43]  -1.4  -0.7  -0.3  -0.7  0.4  1.30       12       13
#> log_lik[44]  -1.3  -0.4  -0.3  -0.6  0.4  0.97       13       13
#> log_lik[45]  -1.2  -0.7  -0.2  -0.7  0.3  1.04        8       13
#> log_lik[46]  -1.1  -0.5  -0.3  -0.6  0.3  1.07       13       13
#> log_lik[47]  -1.6  -0.5  -0.2  -0.7  0.5  0.98       13       13
#> log_lik[48]  -0.9  -0.4  -0.2  -0.5  0.2  1.04       13       13
#> log_lik[49]  -0.8  -0.4  -0.2  -0.4  0.2  0.95       13       13
#> log_lik[50]  -0.6  -0.4  -0.2  -0.4  0.1  0.91       13       13
#> log_lik[51]  -0.7  -0.4  -0.3  -0.4  0.1  0.99       13       13
#> log_lik[52]  -3.7  -2.9  -1.2  -2.7  0.9  0.99       13       13
#> log_lik[53]  -1.6  -0.6  -0.4  -0.8  0.5  0.92       13       13
#> log_lik[54]  -1.7  -0.5  -0.3  -0.8  0.6  0.97       13       13
#> log_lik[55]  -1.3  -0.5  -0.3  -0.7  0.4  0.96       13       13
#> log_lik[56]  -2.2  -1.2  -0.3  -1.1  0.7  0.92       13       13
#> log_lik[57]  -2.6  -1.2  -0.3  -1.2  0.9  0.92       13       13
#> log_lik[58]  -1.1  -0.5  -0.3  -0.6  0.3  1.04        9       13
#> log_lik[59]  -1.5  -0.5  -0.3  -0.6  0.5  1.33        6       13
#> log_lik[60]  -1.1  -0.6  -0.2  -0.6  0.3  1.05       12       13
#> psi[1]        0.4   2.1   5.3   2.4  1.8  0.93       13       13
#> xstar[1,1]   -1.2   0.6   3.6   1.0  1.8  1.09       13       13
#> sigma[1]      0.5   0.5   0.6   0.5  0.0  1.33        6       13
#> nu[1]         4.0  11.9  53.5  20.6 18.6  1.45       13       13
#> lp__        -63.5 -58.5 -52.6 -57.8  3.9  1.11       10       13
#> 
#> For each parameter, Bulk_ESS and Tail_ESS are crude measures of 
#> effective sample size for bulk and tail quantities respectively (an ESS > 100 
#> per chain is considered good), and Rhat is the potential scale reduction 
#> factor on rank normalized split chains (at convergence, Rhat <= 1.05).

#> Warning: Some Pareto k diagnostic values are too high. See help('pareto-k-diagnostic') for details.

#> 
#> SAMPLING FOR MODEL 'dfa' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 6.3e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.63 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: WARNING: There aren't enough warmup iterations to fit the
#> Chain 1:          three stages of adaptation as currently configured.
#> Chain 1:          Reducing each adaptation stage to 15%/75%/10% of
#> Chain 1:          the given number of warmup iterations:
#> Chain 1:            init_buffer = 3
#> Chain 1:            adapt_window = 20
#> Chain 1:            term_buffer = 2
#> Chain 1: 
#> Chain 1: Iteration:  1 / 50 [  2%]  (Warmup)
#> Chain 1: Iteration:  5 / 50 [ 10%]  (Warmup)
#> Chain 1: Iteration: 10 / 50 [ 20%]  (Warmup)
#> Chain 1: Iteration: 15 / 50 [ 30%]  (Warmup)
#> Chain 1: Iteration: 20 / 50 [ 40%]  (Warmup)
#> Chain 1: Iteration: 25 / 50 [ 50%]  (Warmup)
#> Chain 1: Iteration: 26 / 50 [ 52%]  (Sampling)
#> Chain 1: Iteration: 30 / 50 [ 60%]  (Sampling)
#> Chain 1: Iteration: 35 / 50 [ 70%]  (Sampling)
#> Chain 1: Iteration: 40 / 50 [ 80%]  (Sampling)
#> Chain 1: Iteration: 45 / 50 [ 90%]  (Sampling)
#> Chain 1: Iteration: 50 / 50 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.053719 seconds (Warm-up)
#> Chain 1:                1.03643 seconds (Sampling)
#> Chain 1:                1.09015 seconds (Total)
#> Chain 1: 

#> Warning: There were 4 divergent transitions after warmup. See
#> http://mc-stan.org/misc/warnings.html#divergent-transitions-after-warmup
#> to find out why this is a problem and how to eliminate them.

#> Warning: There were 1 chains where the estimated Bayesian Fraction of Missing Information was low. See
#> http://mc-stan.org/misc/warnings.html#bfmi-low

#> Warning: Examine the pairs() plot to diagnose sampling problems

#> Warning: The largest R-hat is NA, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#r-hat

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#tail-ess

#> Inference for the input samples (1 chains: each with iter = 25; warmup = 12):
#> 
#>                Q5   Q50   Q95  Mean   SD  Rhat Bulk_ESS Tail_ESS
#> x[1,1]       -1.6  -0.6   0.1  -0.7  0.6  0.93       10       13
#> x[2,1]        1.7   3.2   4.2   3.0  0.9  0.96       13       13
#> x[1,2]       -1.4  -0.3   0.2  -0.5  0.6  1.07       13       13
#> x[2,2]        1.4   2.9   3.6   2.6  0.9  1.04       13       13
#> x[1,3]       -2.1  -0.7  -0.2  -0.9  0.7  0.94       13       13
#> x[2,3]        0.8   1.3   2.4   1.5  0.6  1.13       13       13
#> x[1,4]       -1.9  -0.6   0.1  -0.7  0.7  1.03        8       13
#> x[2,4]       -0.7   0.2   0.7   0.1  0.5  1.08        9       13
#> x[1,5]       -3.8  -2.3  -0.9  -2.3  1.0  0.97       13       13
#> x[2,5]       -2.2  -1.7  -0.4  -1.5  0.7  0.95       13       13
#> x[1,6]       -2.4  -1.5  -0.2  -1.4  0.8  0.99       13       13
#> x[2,6]       -3.4  -2.7  -1.1  -2.4  0.8  0.93       13       13
#> x[1,7]       -1.8  -1.0   0.3  -0.8  0.8  0.99       13       13
#> x[2,7]       -2.9  -2.5  -1.7  -2.4  0.4  1.13        9       13
#> x[1,8]       -3.9  -1.8  -0.9  -2.0  1.2  1.01       13       13
#> x[2,8]       -2.4  -1.9  -1.1  -1.8  0.5  1.02       13       13
#> x[1,9]       -5.5  -3.5  -1.9  -3.5  1.2  0.92       13       13
#> x[2,9]       -3.0  -1.4  -0.5  -1.6  0.9  1.02       13       13
#> x[1,10]      -5.7  -3.4  -1.9  -3.7  1.4  0.92       13       13
#> x[2,10]      -3.3  -0.9  -0.2  -1.4  1.2  1.02       10       13
#> x[1,11]      -5.4  -3.1  -1.4  -3.4  1.4  1.09       13       13
#> x[2,11]      -1.4   0.2   1.1   0.0  1.0  0.95       10       13
#> x[1,12]      -3.3  -2.1  -0.9  -2.1  0.8  0.98       13       13
#> x[2,12]      -1.2  -0.5   0.4  -0.4  0.6  1.31        6       13
#> x[1,13]      -0.3   0.5   2.0   0.7  0.8  0.92       13       13
#> x[2,13]      -0.6  -0.2   0.7  -0.1  0.4  1.24        7       13
#> x[1,14]       1.4   2.9   4.2   2.8  1.0  1.38       13       13
#> x[2,14]       0.3   1.0   2.3   1.1  0.7  1.30        6       13
#> x[1,15]       2.1   3.9   4.8   3.6  1.0  0.94       13       13
#> x[2,15]       0.4   1.3   2.9   1.5  1.0  1.11       10       13
#> x[1,16]       1.5   3.5   4.7   3.3  1.1  0.97       13       13
#> x[2,16]       0.3   1.3   2.1   1.2  0.7  1.12        9       13
#> x[1,17]       1.5   3.2   5.5   3.3  1.6  1.00       13       13
#> x[2,17]      -0.2   0.3   2.7   0.8  1.0  1.37        5       13
#> x[1,18]       1.4   3.3   5.3   3.4  1.3  1.18       13       13
#> x[2,18]      -1.6  -1.0   0.7  -0.7  0.8  1.08       11       13
#> x[1,19]       1.1   2.4   3.7   2.5  1.0  0.93       13       13
#> x[2,19]      -1.6  -1.0  -0.2  -0.9  0.5  0.97       13       13
#> x[1,20]       1.0   2.6   3.8   2.6  1.0  1.09       13       13
#> x[2,20]      -0.2   1.2   2.1   1.0  0.8  0.99       12       13
#> Z[1,1]        0.3   0.4   0.9   0.4  0.3  1.21       13       13
#> Z[2,1]       -0.1   0.2   0.3   0.1  0.1  0.93       13       13
#> Z[3,1]        0.0   0.2   0.4   0.2  0.2  1.07       13       13
#> Z[1,2]        0.0   0.0   0.0   0.0  0.0  1.00       13       13
#> Z[2,2]        0.4   0.4   0.7   0.5  0.1  1.03       13       13
#> Z[3,2]        0.4   0.5   0.7   0.5  0.1  1.00       13       13
#> log_lik[1]   -0.7   0.1   0.3  -0.1  0.4  1.01        7       13
#> log_lik[2]   -0.6   0.0   0.3   0.0  0.4  0.95       13       13
#> log_lik[3]   -1.1   0.1   0.2  -0.1  0.5  1.12       10       13
#> log_lik[4]   -1.8   0.1   0.2  -0.3  0.7  1.12        7       13
#> log_lik[5]   -0.1   0.1   0.4   0.2  0.2  1.03       13       13
#> log_lik[6]   -0.6   0.1   0.4   0.1  0.3  0.96       13       13
#> log_lik[7]   -2.6  -1.4   0.3  -1.2  1.1  0.94       13       13
#> log_lik[8]   -0.7   0.0   0.4   0.0  0.4  0.95       13       13
#> log_lik[9]   -0.3   0.2   0.4   0.1  0.2  0.96       13       13
#> log_lik[10]  -4.8  -2.3  -0.5  -2.5  1.6  1.15        6       13
#> log_lik[11]  -0.4   0.0   0.3   0.0  0.2  1.39        4       13
#> log_lik[12]  -0.4   0.1   0.3   0.0  0.3  0.96        9       13
#> log_lik[13]  -3.7  -0.7   0.0  -1.2  1.4  0.95       13       13
#> log_lik[14]  -0.2   0.2   0.4   0.1  0.2  1.27        6       13
#> log_lik[15]  -0.7   0.0   0.3  -0.1  0.4  1.09       13       13
#> log_lik[16]  -0.6   0.0   0.4   0.0  0.3  1.04       10       13
#> log_lik[17]  -0.6   0.1   0.3   0.0  0.3  0.98       13       13
#> log_lik[18]  -1.2   0.1   0.3  -0.1  0.5  1.45        5       13
#> log_lik[19]  -2.0  -1.3   0.3  -0.9  0.9  1.07       13       13
#> log_lik[20]  -0.7  -0.2   0.2  -0.2  0.3  1.24       12       13
#> log_lik[21]  -0.4   0.1   0.2   0.0  0.2  0.93       13       13
#> log_lik[22]  -0.6   0.1   0.3   0.0  0.3  1.24        7       13
#> log_lik[23]  -0.2   0.0   0.3   0.0  0.2  1.16        8       13
#> log_lik[24]  -0.1   0.1   0.3   0.1  0.2  1.00       11       13
#> log_lik[25]  -0.8   0.0   0.2  -0.1  0.4  0.94       13       13
#> log_lik[26]  -0.2   0.2   0.3   0.1  0.2  1.12       13       13
#> log_lik[27]  -0.1   0.2   0.3   0.1  0.1  1.05       10       13
#> log_lik[28]  -1.1   0.2   0.4  -0.1  0.6  1.71        4       13
#> log_lik[29]  -0.6   0.0   0.2  -0.1  0.3  1.04       12       13
#> log_lik[30]  -1.6  -0.2   0.2  -0.4  0.8  0.92       13       13
#> log_lik[31]  -1.1   0.2   0.2  -0.2  0.5  1.06       11       13
#> log_lik[32]  -0.9  -0.3   0.3  -0.3  0.4  0.94       13       13
#> log_lik[33]  -0.4   0.1   0.4   0.0  0.3  1.31       10       13
#> log_lik[34]  -2.1  -0.2   0.2  -0.5  0.9  1.13        8       13
#> log_lik[35]  -1.7  -0.1   0.3  -0.4  0.7  1.32        6       13
#> log_lik[36]  -0.6   0.1   0.3   0.0  0.3  1.00       13       13
#> log_lik[37]  -2.1  -0.2   0.2  -0.6  0.9  0.94       13       13
#> log_lik[38]  -0.9   0.2   0.3   0.0  0.4  1.87        4       13
#> log_lik[39]  -0.3   0.2   0.4   0.1  0.2  1.87        4       13
#> log_lik[40]  -0.7   0.1   0.3   0.0  0.4  1.01       12       13
#> log_lik[41]  -0.1   0.1   0.4   0.2  0.2  1.71        4       13
#> log_lik[42]  -1.4  -0.9   0.1  -0.7  0.5  1.02       13       13
#> log_lik[43]  -0.8   0.1   0.4  -0.1  0.5  0.92       13       13
#> log_lik[44]  -1.1  -0.1   0.2  -0.2  0.4  1.19       13       13
#> log_lik[45]  -1.1  -0.1   0.3  -0.3  0.5  1.07       13       13
#> log_lik[46]  -1.3   0.1   0.4  -0.1  0.6  1.12        8       13
#> log_lik[47]  -1.3   0.1   0.3  -0.3  0.6  1.00       11       13
#> log_lik[48]  -0.9   0.0   0.4  -0.1  0.4  1.19       13       13
#> log_lik[49]  -2.3  -0.3   0.1  -0.7  0.9  0.92       13       13
#> log_lik[50]  -1.2   0.1   0.3  -0.1  0.6  1.31        6       13
#> log_lik[51]  -0.5   0.1   0.3   0.0  0.3  1.24        6       13
#> log_lik[52]  -1.3  -0.2   0.2  -0.4  0.6  1.14        8       13
#> log_lik[53]  -0.1   0.2   0.4   0.1  0.2  0.98       13       13
#> log_lik[54]  -0.1   0.2   0.4   0.2  0.2  0.98       13       13
#> log_lik[55]  -1.2  -0.1   0.1  -0.3  0.5  0.99       13       13
#> log_lik[56]  -1.0   0.0   0.2  -0.2  0.4  0.99       13       13
#> log_lik[57]  -1.3   0.0   0.2  -0.2  0.6  0.99       13       13
#> log_lik[58]  -0.9   0.0   0.2  -0.1  0.4  0.92       13       13
#> log_lik[59]  -1.4   0.1   0.4  -0.3  0.7  1.19       13       13
#> log_lik[60]  -0.8  -0.1   0.3  -0.2  0.4  0.93       13       13
#> psi[1]        0.8   2.1   6.7   2.8  2.2  0.93       13       13
#> psi[2]        0.9   1.8   4.0   2.0  1.1  1.71        4       13
#> xstar[1,1]    1.2   2.7   4.8   2.9  1.4  1.21       13       13
#> xstar[2,1]   -3.0  -0.4   1.0  -0.5  1.6  0.97       13       13
#> sigma[1]      0.3   0.3   0.3   0.3  0.0  1.30        7       13
#> nu[1]        13.9  20.9  44.1  22.6 10.6  0.93       13       13
#> lp__        -45.3 -33.2 -23.4 -34.1  7.6  1.39        5       13
#> 
#> For each parameter, Bulk_ESS and Tail_ESS are crude measures of 
#> effective sample size for bulk and tail quantities respectively (an ESS > 100 
#> per chain is considered good), and Rhat is the potential scale reduction 
#> factor on rank normalized split chains (at convergence, Rhat <= 1.05).

#> Warning: Some Pareto k diagnostic values are too high. See help('pareto-k-diagnostic') for details.

m$summary

#>   model num_trends    looic   cor     error converge
#> 1     1          1 107.1559 equal student-t    FALSE
#> 2     2          2 235.8225 equal student-t    FALSE

m$best_model

#> NULL

# }