Fit models with differing numbers of regimes to trend data

Fit models with differing numbers of regimes to trend data

fit_regimes(
  y,
  sds = NULL,
  n_regimes = 2,
  iter = 2000,
  thin = 1,
  chains = 1,
  ...
)

Arguments

y
Data, time series or trend from fitted DFA model.

sds
Optional time series of standard deviations of estimates. If passed in, residual variance not estimated. Defaults to NULL.

n_regimes
Number of regimes to evaluate, defaults 2

iter
MCMC iterations, defaults to 2000.

thin
MCMC thinning rate, defaults to 1.

chains
MCMC chains, defaults to 1 (note that running multiple chains may result in a label switching problem where the regimes are identified with different IDs across chains).

...
Other parameters to pass to rstan::sampling().

Examples

data(Nile)
fit_regimes(log(Nile), iter = 50, n_regimes = 1)

#> 
#> SAMPLING FOR MODEL 'regime_1' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 9e-06 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.09 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: WARNING: There aren't enough warmup iterations to fit the
#> Chain 1:          three stages of adaptation as currently configured.
#> Chain 1:          Reducing each adaptation stage to 15%/75%/10% of
#> Chain 1:          the given number of warmup iterations:
#> Chain 1:            init_buffer = 3
#> Chain 1:            adapt_window = 20
#> Chain 1:            term_buffer = 2
#> Chain 1: 
#> Chain 1: Iteration:  1 / 50 [  2%]  (Warmup)
#> Chain 1: Iteration:  5 / 50 [ 10%]  (Warmup)
#> Chain 1: Iteration: 10 / 50 [ 20%]  (Warmup)
#> Chain 1: Iteration: 15 / 50 [ 30%]  (Warmup)
#> Chain 1: Iteration: 20 / 50 [ 40%]  (Warmup)
#> Chain 1: Iteration: 25 / 50 [ 50%]  (Warmup)
#> Chain 1: Iteration: 26 / 50 [ 52%]  (Sampling)
#> Chain 1: Iteration: 30 / 50 [ 60%]  (Sampling)
#> Chain 1: Iteration: 35 / 50 [ 70%]  (Sampling)
#> Chain 1: Iteration: 40 / 50 [ 80%]  (Sampling)
#> Chain 1: Iteration: 45 / 50 [ 90%]  (Sampling)
#> Chain 1: Iteration: 50 / 50 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.000908 seconds (Warm-up)
#> Chain 1:                0.000791 seconds (Sampling)
#> Chain 1:                0.001699 seconds (Total)
#> Chain 1: 

#> Warning: The largest R-hat is 1.52, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#r-hat

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#tail-ess

#> Warning: Some Pareto k diagnostic values are too high. See help('pareto-k-diagnostic') for details.

#> $model
#> Inference for Stan model: regime_1.
#> 1 chains, each with iter=50; warmup=25; thin=1; 
#> post-warmup draws per chain=25, total post-warmup draws=25.
#> 
#>                mean se_mean   sd   2.5%    25%    50%    75%  97.5% n_eff Rhat
#> mu_k           6.81    0.00 0.02   6.77   6.79   6.81   6.82   6.83    25 0.96
#> sigma_k        0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[1]      0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[2]      0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[3]      0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[4]      0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[5]      0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[6]      0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[7]      0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[8]      0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[9]      0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[10]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[11]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[12]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[13]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[14]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[15]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[16]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[17]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[18]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[19]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[20]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[21]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[22]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[23]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[24]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[25]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[26]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[27]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[28]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[29]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[30]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[31]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[32]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[33]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[34]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[35]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[36]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[37]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[38]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[39]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[40]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[41]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[42]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[43]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[44]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[45]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[46]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[47]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[48]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[49]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[50]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[51]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[52]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[53]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[54]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[55]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[56]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[57]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[58]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[59]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[60]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[61]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[62]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[63]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[64]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[65]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[66]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[67]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[68]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[69]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[70]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[71]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[72]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[73]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[74]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[75]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[76]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[77]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[78]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[79]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[80]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[81]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[82]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[83]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[84]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[85]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[86]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[87]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[88]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[89]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[90]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[91]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[92]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[93]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[94]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[95]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[96]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[97]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[98]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[99]     0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> sigmas[100]    0.18    0.01 0.02   0.16   0.16   0.19   0.19   0.21     5 1.40
#> log_lik[1]     0.07    0.02 0.13  -0.14  -0.02   0.07   0.18   0.24    35 1.02
#> log_lik[2]    -0.18    0.03 0.15  -0.43  -0.29  -0.18  -0.05   0.04    23 1.14
#> log_lik[3]     0.72    0.04 0.11   0.56   0.64   0.71   0.82   0.88     6 1.19
#> log_lik[4]    -0.54    0.06 0.20  -0.84  -0.72  -0.53  -0.35  -0.24    10 1.29
#> log_lik[5]    -0.18    0.03 0.15  -0.43  -0.29  -0.18  -0.05   0.04    23 1.14
#> log_lik[6]    -0.18    0.03 0.15  -0.43  -0.29  -0.18  -0.05   0.04    23 1.14
#> log_lik[7]     0.61    0.02 0.07   0.48   0.55   0.61   0.67   0.72    10 1.25
#> log_lik[8]    -0.69    0.08 0.22  -1.03  -0.88  -0.69  -0.48  -0.36     8 1.34
#> log_lik[9]    -1.91    0.18 0.42  -2.59  -2.13  -1.98  -1.53  -1.23     5 1.53
#> log_lik[10]   -0.05    0.02 0.14  -0.28  -0.13  -0.05   0.05   0.14    32 1.07
#> log_lik[11]    0.64    0.04 0.11   0.48   0.55   0.64   0.73   0.81     8 1.09
#> log_lik[12]    0.77    0.04 0.10   0.61   0.70   0.74   0.87   0.91     5 1.28
#> log_lik[13]    0.13    0.02 0.12  -0.08   0.04   0.13   0.25   0.29    35 1.00
#> log_lik[14]    0.64    0.04 0.11   0.48   0.56   0.65   0.73   0.81     8 1.09
#> log_lik[15]    0.56    0.03 0.11   0.38   0.47   0.56   0.62   0.72    10 1.03
#> log_lik[16]    0.73    0.04 0.11   0.57   0.65   0.71   0.83   0.89     6 1.20
#> log_lik[17]   -0.32    0.04 0.17  -0.59  -0.46  -0.32  -0.16  -0.07    15 1.20
#> log_lik[18]    0.55    0.02 0.08   0.43   0.49   0.56   0.61   0.66    20 1.12
#> log_lik[19]    0.73    0.04 0.10   0.57   0.65   0.71   0.83   0.89     6 1.20
#> log_lik[20]   -0.05    0.02 0.14  -0.28  -0.13  -0.05   0.05   0.14    32 1.07
#> log_lik[21]    0.18    0.02 0.12  -0.02   0.10   0.18   0.29   0.33    31 0.98
#> log_lik[22]   -0.54    0.06 0.20  -0.84  -0.72  -0.53  -0.35  -0.24    10 1.29
#> log_lik[23]   -0.12    0.03 0.15  -0.36  -0.21  -0.12   0.00   0.09    29 1.10
#> log_lik[24]   -0.85    0.09 0.24  -1.23  -1.05  -0.88  -0.62  -0.48     7 1.39
#> log_lik[25]   -0.93    0.10 0.26  -1.34  -1.13  -0.96  -0.69  -0.53     7 1.41
#> log_lik[26]   -0.61    0.07 0.21  -0.93  -0.80  -0.60  -0.42  -0.30     9 1.32
#> log_lik[27]    0.52    0.03 0.11   0.34   0.43   0.53   0.58   0.69    11 1.00
#> log_lik[28]    0.18    0.02 0.12  -0.02   0.10   0.18   0.29   0.33    31 0.98
#> log_lik[29]    0.41    0.02 0.09   0.29   0.33   0.39   0.49   0.55    35 0.98
#> log_lik[30]    0.70    0.03 0.08   0.56   0.64   0.71   0.74   0.83     6 1.43
#> log_lik[31]    0.77    0.04 0.09   0.62   0.71   0.73   0.86   0.90     5 1.45
#> log_lik[32]   -0.30    0.08 0.23  -0.69  -0.43  -0.27  -0.10   0.05     9 1.05
#> log_lik[33]    0.76    0.04 0.10   0.61   0.68   0.73   0.86   0.91     6 1.26
#> log_lik[34]    0.68    0.03 0.08   0.54   0.62   0.70   0.72   0.80     6 1.40
#> log_lik[35]   -0.22    0.07 0.21  -0.58  -0.33  -0.20  -0.04   0.11    10 1.04
#> log_lik[36]    0.78    0.04 0.10   0.64   0.71   0.75   0.90   0.92     5 1.34
#> log_lik[37]   -0.32    0.08 0.23  -0.73  -0.46  -0.30  -0.12   0.03     9 1.05
#> log_lik[38]    0.56    0.03 0.11   0.38   0.47   0.56   0.62   0.72    10 1.03
#> log_lik[39]    0.44    0.03 0.11   0.24   0.34   0.46   0.51   0.60    15 0.97
#> log_lik[40]    0.71    0.04 0.11   0.55   0.63   0.70   0.81   0.87     6 1.17
#> log_lik[41]    0.68    0.03 0.08   0.54   0.62   0.69   0.71   0.80     6 1.39
#> log_lik[42]    0.04    0.04 0.16  -0.23  -0.07   0.02   0.17   0.29    13 0.99
#> log_lik[43]   -6.50    0.66 1.50  -8.88  -7.78  -5.99  -5.32  -4.50     5 1.29
#> log_lik[44]    0.65    0.03 0.08   0.52   0.59   0.66   0.70   0.77     7 1.34
#> log_lik[45]   -0.21    0.07 0.21  -0.57  -0.32  -0.19  -0.03   0.12    10 1.04
#> log_lik[46]    0.07    0.02 0.13  -0.14  -0.02   0.07   0.18   0.24    35 1.02
#> log_lik[47]    0.18    0.02 0.12  -0.02   0.10   0.18   0.29   0.33    31 0.98
#> log_lik[48]    0.68    0.03 0.08   0.54   0.62   0.69   0.72   0.80     6 1.39
#> log_lik[49]    0.35    0.02 0.10   0.21   0.27   0.33   0.44   0.50    30 0.96
#> log_lik[50]    0.64    0.03 0.07   0.51   0.58   0.64   0.69   0.76     8 1.32
#> log_lik[51]    0.37    0.02 0.10   0.24   0.30   0.35   0.46   0.52    33 0.97
#> log_lik[52]    0.71    0.03 0.08   0.57   0.66   0.72   0.76   0.84     5 1.45
#> log_lik[53]    0.75    0.04 0.09   0.61   0.70   0.73   0.83   0.89     5 1.47
#> log_lik[54]    0.75    0.04 0.08   0.61   0.70   0.73   0.82   0.88     5 1.47
#> log_lik[55]   -0.25    0.07 0.22  -0.63  -0.37  -0.23  -0.06   0.08    10 1.04
#> log_lik[56]    0.71    0.03 0.08   0.57   0.66   0.72   0.76   0.84     5 1.45
#> log_lik[57]    0.20    0.03 0.13  -0.01   0.11   0.18   0.31   0.40    18 0.96
#> log_lik[58]    0.53    0.02 0.08   0.42   0.47   0.55   0.60   0.65    24 1.10
#> log_lik[59]    0.48    0.03 0.11   0.29   0.39   0.49   0.55   0.64    13 0.99
#> log_lik[60]    0.31    0.02 0.11   0.16   0.23   0.29   0.41   0.48    26 0.96
#> log_lik[61]    0.45    0.01 0.09   0.34   0.37   0.45   0.52   0.58    35 1.01
#> log_lik[62]    0.75    0.04 0.09   0.61   0.70   0.73   0.83   0.89     5 1.47
#> log_lik[63]    0.71    0.03 0.08   0.57   0.66   0.72   0.76   0.84     5 1.45
#> log_lik[64]    0.76    0.04 0.10   0.60   0.68   0.73   0.85   0.91     6 1.25
#> log_lik[65]    0.67    0.04 0.11   0.51   0.59   0.67   0.76   0.84     7 1.12
#> log_lik[66]    0.78    0.04 0.10   0.64   0.72   0.75   0.89   0.92     5 1.40
#> log_lik[67]    0.64    0.03 0.07   0.51   0.59   0.65   0.70   0.76     8 1.32
#> log_lik[68]    0.59    0.04 0.11   0.42   0.51   0.59   0.67   0.76     9 1.05
#> log_lik[69]    0.39    0.02 0.09   0.27   0.31   0.37   0.47   0.54    35 0.97
#> log_lik[70]   -0.53    0.10 0.28  -1.00  -0.71  -0.52  -0.28  -0.12     8 1.08
#> log_lik[71]   -0.92    0.13 0.36  -1.54  -1.19  -0.89  -0.60  -0.42     7 1.12
#> log_lik[72]    0.72    0.03 0.08   0.57   0.66   0.72   0.76   0.84     5 1.45
#> log_lik[73]    0.61    0.02 0.07   0.48   0.55   0.61   0.67   0.71    10 1.24
#> log_lik[74]    0.18    0.03 0.13  -0.03   0.08   0.16   0.29   0.39    17 0.97
#> log_lik[75]    0.56    0.02 0.08   0.44   0.50   0.57   0.62   0.67    18 1.14
#> log_lik[76]    0.48    0.03 0.11   0.29   0.39   0.49   0.55   0.64    13 0.99
#> log_lik[77]    0.75    0.04 0.08   0.60   0.69   0.73   0.82   0.88     5 1.47
#> log_lik[78]    0.77    0.04 0.09   0.62   0.71   0.73   0.86   0.90     5 1.45
#> log_lik[79]    0.72    0.04 0.08   0.58   0.67   0.72   0.77   0.85     5 1.46
#> log_lik[80]    0.78    0.04 0.09   0.64   0.72   0.75   0.88   0.92     5 1.42
#> log_lik[81]    0.20    0.03 0.13  -0.01   0.11   0.18   0.31   0.40    18 0.96
#> log_lik[82]    0.24    0.03 0.12   0.05   0.16   0.23   0.35   0.43    20 0.96
#> log_lik[83]    0.70    0.03 0.08   0.55   0.64   0.70   0.73   0.82     6 1.42
#> log_lik[84]    0.44    0.03 0.11   0.24   0.34   0.46   0.51   0.60    15 0.97
#> log_lik[85]    0.78    0.04 0.10   0.64   0.71   0.75   0.89   0.92     5 1.33
#> log_lik[86]    0.67    0.04 0.11   0.50   0.58   0.66   0.76   0.83     7 1.12
#> log_lik[87]    0.54    0.02 0.08   0.42   0.48   0.56   0.60   0.65    23 1.11
#> log_lik[88]    0.78    0.04 0.10   0.63   0.71   0.75   0.88   0.92     5 1.32
#> log_lik[89]    0.70    0.04 0.11   0.53   0.61   0.68   0.79   0.86     7 1.15
#> log_lik[90]    0.62    0.02 0.07   0.49   0.56   0.62   0.68   0.73     9 1.26
#> log_lik[91]    0.56    0.03 0.11   0.38   0.47   0.56   0.62   0.72    10 1.03
#> log_lik[92]    0.78    0.04 0.10   0.64   0.71   0.74   0.90   0.92     5 1.37
#> log_lik[93]    0.78    0.04 0.10   0.64   0.72   0.74   0.89   0.92     5 1.39
#> log_lik[94]   -0.25    0.04 0.16  -0.51  -0.37  -0.25  -0.11  -0.01    18 1.17
#> log_lik[95]    0.78    0.04 0.10   0.64   0.71   0.74   0.90   0.92     5 1.35
#> log_lik[96]    0.21    0.03 0.13   0.02   0.13   0.20   0.32   0.41    19 0.96
#> log_lik[97]    0.78    0.04 0.10   0.64   0.71   0.75   0.89   0.92     5 1.33
#> log_lik[98]   -0.04    0.05 0.17  -0.33  -0.15  -0.04   0.11   0.24    12 1.01
#> log_lik[99]   -0.08    0.05 0.18  -0.39  -0.19  -0.08   0.08   0.21    11 1.01
#> log_lik[100]   0.16    0.03 0.13  -0.05   0.06   0.15   0.27   0.38    17 0.97
#> lp__         113.34    0.44 1.19 111.43 112.13 113.48 114.37 114.92     7 1.50
#> 
#> Samples were drawn using NUTS(diag_e) at Thu Apr 29 15:25:37 2021.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).
#> 
#> $y
#> Time Series:
#> Start = 1871 
#> End = 1970 
#> Frequency = 1 
#>   [1] 7.021084 7.056175 6.870053 7.098376 7.056175 7.056175 6.700731 7.114769
#>   [9] 7.222566 7.038784 6.902743 6.840547 7.012115 6.901737 6.927558 6.866933
#>  [17] 7.073270 6.683361 6.864848 7.038784 7.003065 7.098376 7.047517 7.130899
#>  [25] 7.138867 7.106606 6.937314 7.003065 6.651572 6.733402 6.773080 6.542472
#>  [33] 6.845880 6.725034 6.552508 6.820016 6.539586 6.927558 6.956545 6.876265
#>  [41] 6.722630 6.587550 6.122493 6.714171 6.553933 7.021084 7.003065 6.723832
#>  [49] 6.638568 6.710523 6.643790 6.739337 6.761573 6.759255 6.548219 6.739337
#>  [57] 6.612041 6.679599 6.946976 6.632002 6.660575 6.762730 6.739337 6.850126
#>  [65] 6.891626 6.799056 6.711740 6.917706 6.647688 6.516193 6.475433 6.740519
#>  [73] 6.699500 6.609349 6.685861 6.946976 6.756932 6.773080 6.742881 6.791221
#>  [81] 6.612041 6.618739 6.731018 6.956545 6.822197 6.893656 6.680855 6.827629
#>  [89] 6.882437 6.703188 6.927558 6.809039 6.803505 7.064759 6.815640 6.614726
#>  [97] 6.823286 6.576470 6.570883 6.606650
#> 
#> $looic
#> [1] -46.52446
#>